Network.AWS.CloudFront.CreateCloudFrontOriginAccessIdentity

Description

Create a new origin access identity.

http://docs.aws.amazon.com/AmazonCloudFront/latest/APIReference/CreateCloudFrontOriginAccessIdentity.html

Request

data CreateCloudFrontOriginAccessIdentity

Request constructor

createCloudFrontOriginAccessIdentity

Arguments

:: CloudFrontOriginAccessIdentityConfig
ccfoaiCloudFrontOriginAccessIdentityConfig

-> CreateCloudFrontOriginAccessIdentity

CreateCloudFrontOriginAccessIdentity constructor.

The fields accessible through corresponding lenses are:

ccfoaiCloudFrontOriginAccessIdentityConfig :: CloudFrontOriginAccessIdentityConfig

Request lenses

ccfoaiCloudFrontOriginAccessIdentityConfig :: Lens' CreateCloudFrontOriginAccessIdentity CloudFrontOriginAccessIdentityConfig

The origin access identity's configuration information.

Response

data CreateCloudFrontOriginAccessIdentityResponse

Response constructor

createCloudFrontOriginAccessIdentityResponse :: CreateCloudFrontOriginAccessIdentityResponse

CreateCloudFrontOriginAccessIdentityResponse constructor.

The fields accessible through corresponding lenses are:

ccfoairCloudFrontOriginAccessIdentity :: Maybe CloudFrontOriginAccessIdentity

ccfoairETag :: Maybe Text

ccfoairLocation :: Maybe Text

Response lenses

ccfoairCloudFrontOriginAccessIdentity :: Lens' CreateCloudFrontOriginAccessIdentityResponse (Maybe CloudFrontOriginAccessIdentity)

The origin access identity's information.

ccfoairETag :: Lens' CreateCloudFrontOriginAccessIdentityResponse (Maybe Text)

The current version of the origin access identity created.

ccfoairLocation :: Lens' CreateCloudFrontOriginAccessIdentityResponse (Maybe Text)

The fully qualified URI of the new origin access identity just created. For example: https://cloudfront.amazonaws.com/2010-11-01/origin-access-identity/cloudfront/E74FTE3AJFJ256A.