Network.AWS.CloudFront.CreateDistribution

Description

Creates a new web distribution. Send a POST request to the /CloudFront API version /distribution /distribution ID resource.

Creating a Request

createDistribution

Arguments

:: DistributionConfig
cdDistributionConfig

-> CreateDistribution

Creates a value of CreateDistribution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdDistributionConfig - The distribution's configuration information.

data CreateDistribution

The request to create a new distribution.

See: createDistribution smart constructor.

Request Lenses

cdDistributionConfig :: Lens' CreateDistribution DistributionConfig

The distribution's configuration information.

Destructuring the Response

createDistributionResponse

Arguments

:: Int
cdrsResponseStatus

-> CreateDistributionResponse

Creates a value of CreateDistributionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdrsETag - The current version of the distribution created.

cdrsDistribution - The distribution's information.

cdrsLocation - The fully qualified URI of the new distribution resource just created. For example: https://cloudfront.amazonaws.com/2010-11-01/distribution/EDFDVBD632BHDS5 .

cdrsResponseStatus - -- | The response status code.

data CreateDistributionResponse

The returned result of the corresponding request.

See: createDistributionResponse smart constructor.

Response Lenses

cdrsETag :: Lens' CreateDistributionResponse (Maybe Text)

The current version of the distribution created.

cdrsDistribution :: Lens' CreateDistributionResponse (Maybe Distribution)

The distribution's information.

cdrsLocation :: Lens' CreateDistributionResponse (Maybe Text)

The fully qualified URI of the new distribution resource just created. For example: https://cloudfront.amazonaws.com/2010-11-01/distribution/EDFDVBD632BHDS5 .

cdrsResponseStatus :: Lens' CreateDistributionResponse Int

- | The response status code.