Network.AWS.CloudFront.CreateDistributionWithTags

Description

Create a new distribution with tags.

Creating a Request

createDistributionWithTags

Arguments

:: DistributionConfigWithTags
cdwtDistributionConfigWithTags

-> CreateDistributionWithTags

Creates a value of CreateDistributionWithTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdwtDistributionConfigWithTags - The distribution's configuration information.

data CreateDistributionWithTags

The request to create a new distribution with tags.

See: createDistributionWithTags smart constructor.

Request Lenses

cdwtDistributionConfigWithTags :: Lens' CreateDistributionWithTags DistributionConfigWithTags

The distribution's configuration information.

Destructuring the Response

createDistributionWithTagsResponse

Arguments

:: Int
cdwtrsResponseStatus

-> CreateDistributionWithTagsResponse

Creates a value of CreateDistributionWithTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdwtrsETag - The current version of the distribution created.

cdwtrsDistribution - The distribution's information.

cdwtrsLocation - The fully qualified URI of the new distribution resource just created. For example: https://cloudfront.amazonaws.com/2010-11-01/distribution/EDFDVBD632BHDS5 .

cdwtrsResponseStatus - -- | The response status code.

data CreateDistributionWithTagsResponse

The returned result of the corresponding request.

See: createDistributionWithTagsResponse smart constructor.

Response Lenses

cdwtrsETag :: Lens' CreateDistributionWithTagsResponse (Maybe Text)

The current version of the distribution created.

cdwtrsDistribution :: Lens' CreateDistributionWithTagsResponse (Maybe Distribution)

The distribution's information.

cdwtrsLocation :: Lens' CreateDistributionWithTagsResponse (Maybe Text)

The fully qualified URI of the new distribution resource just created. For example: https://cloudfront.amazonaws.com/2010-11-01/distribution/EDFDVBD632BHDS5 .

cdwtrsResponseStatus :: Lens' CreateDistributionWithTagsResponse Int

- | The response status code.