Network.AWS.CloudFront.CreateInvalidation

Description

Create a new invalidation.

Creating a Request

createInvalidation

Arguments

:: Text
ciDistributionId

-> InvalidationBatch
ciInvalidationBatch

-> CreateInvalidation

Creates a value of CreateInvalidation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciDistributionId - The distribution's id.

ciInvalidationBatch - The batch information for the invalidation.

data CreateInvalidation

The request to create an invalidation.

See: createInvalidation smart constructor.

Request Lenses

ciDistributionId :: Lens' CreateInvalidation Text

The distribution's id.

ciInvalidationBatch :: Lens' CreateInvalidation InvalidationBatch

The batch information for the invalidation.

Destructuring the Response

createInvalidationResponse

Arguments

:: Int
cirsResponseStatus

-> CreateInvalidationResponse

Creates a value of CreateInvalidationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cirsInvalidation - The invalidation's information.

cirsLocation - The fully qualified URI of the distribution and invalidation batch request, including the Invalidation ID .

cirsResponseStatus - -- | The response status code.

data CreateInvalidationResponse

The returned result of the corresponding request.

See: createInvalidationResponse smart constructor.

Response Lenses

cirsInvalidation :: Lens' CreateInvalidationResponse (Maybe Invalidation)

The invalidation's information.

cirsLocation :: Lens' CreateInvalidationResponse (Maybe Text)

The fully qualified URI of the distribution and invalidation batch request, including the Invalidation ID .

cirsResponseStatus :: Lens' CreateInvalidationResponse Int

- | The response status code.