Network.AWS.CloudFront.CreateStreamingDistribution

Description

Creates a new RMTP distribution. An RTMP distribution is similar to a web distribution, but an RTMP distribution streams media files using the Adobe Real-Time Messaging Protocol (RTMP) instead of serving files using HTTP.

To create a new web distribution, submit a POST request to the CloudFront API version distribution resource. The request body must include a document with a StreamingDistributionConfig/ element. The response echoes the StreamingDistributionConfig element and returns other information about the RTMP distribution.

To get the status of your request, use the GET StreamingDistribution API action. When the value of Enabled is true and the value of Status is Deployed , your distribution is ready. A distribution usually deploys in less than 15 minutes.

For more information about web distributions, see Working with RTMP Distributions in the Amazon CloudFront Developer Guide .

Important: Beginning with the 2012-05-05 version of the CloudFront API, we made substantial changes to the format of the XML document that you include in the request body when you create or update a web distribution or an RTMP distribution, and when you invalidate objects. With previous versions of the API, we discovered that it was too easy to accidentally delete one or more values for an element that accepts multiple values, for example, CNAMEs and trusted signers. Our changes for the 2012-05-05 release are intended to prevent these accidental deletions and to notify you when there's a mismatch between the number of values you say you're specifying in the Quantity element and the number of values specified.

Creating a Request

createStreamingDistribution

Arguments

:: StreamingDistributionConfig
csdStreamingDistributionConfig

-> CreateStreamingDistribution

Creates a value of CreateStreamingDistribution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csdStreamingDistributionConfig - The streaming distribution's configuration information.

data CreateStreamingDistribution

The request to create a new streaming distribution.

See: createStreamingDistribution smart constructor.

Request Lenses

csdStreamingDistributionConfig :: Lens' CreateStreamingDistribution StreamingDistributionConfig

The streaming distribution's configuration information.

Destructuring the Response

createStreamingDistributionResponse

Arguments

:: Int
csdrsResponseStatus

-> CreateStreamingDistributionResponse

Creates a value of CreateStreamingDistributionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csdrsETag - The current version of the streaming distribution created.

csdrsLocation - The fully qualified URI of the new streaming distribution resource just created. For example: https://cloudfront.amazonaws.com/2010-11-01/streaming-distribution/EGTXBD79H29TRA8 .

csdrsStreamingDistribution - The streaming distribution's information.

csdrsResponseStatus - -- | The response status code.

data CreateStreamingDistributionResponse

The returned result of the corresponding request.

See: createStreamingDistributionResponse smart constructor.

Response Lenses

csdrsETag :: Lens' CreateStreamingDistributionResponse (Maybe Text)

The current version of the streaming distribution created.

csdrsLocation :: Lens' CreateStreamingDistributionResponse (Maybe Text)

The fully qualified URI of the new streaming distribution resource just created. For example: https://cloudfront.amazonaws.com/2010-11-01/streaming-distribution/EGTXBD79H29TRA8 .

csdrsStreamingDistribution :: Lens' CreateStreamingDistributionResponse (Maybe StreamingDistribution)

The streaming distribution's information.

csdrsResponseStatus :: Lens' CreateStreamingDistributionResponse Int

- | The response status code.