Network.AWS.CloudFront.DeleteCloudFrontOriginAccessIdentity

Description

Delete an origin access identity.

Creating a Request

deleteCloudFrontOriginAccessIdentity

Arguments

:: Text
dcfoaiId

-> DeleteCloudFrontOriginAccessIdentity

Creates a value of DeleteCloudFrontOriginAccessIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcfoaiIfMatch - The value of the ETag header you received from a previous GET or PUT request. For example: E2QWRUHAPOMQZL .

dcfoaiId - The origin access identity's ID.

data DeleteCloudFrontOriginAccessIdentity

Deletes a origin access identity.

See: deleteCloudFrontOriginAccessIdentity smart constructor.

Request Lenses

dcfoaiIfMatch :: Lens' DeleteCloudFrontOriginAccessIdentity (Maybe Text)

The value of the ETag header you received from a previous GET or PUT request. For example: E2QWRUHAPOMQZL .

dcfoaiId :: Lens' DeleteCloudFrontOriginAccessIdentity Text

The origin access identity's ID.

Destructuring the Response

deleteCloudFrontOriginAccessIdentityResponse :: DeleteCloudFrontOriginAccessIdentityResponse

Creates a value of DeleteCloudFrontOriginAccessIdentityResponse with the minimum fields required to make a request.

data DeleteCloudFrontOriginAccessIdentityResponse

See: deleteCloudFrontOriginAccessIdentityResponse smart constructor.