Network.AWS.CloudFront.GetCloudFrontOriginAccessIdentity

Description

Get the information about an origin access identity.

Creating a Request

getCloudFrontOriginAccessIdentity

Arguments

:: Text
gcfoaiId

-> GetCloudFrontOriginAccessIdentity

Creates a value of GetCloudFrontOriginAccessIdentity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcfoaiId - The identity's ID.

data GetCloudFrontOriginAccessIdentity

The request to get an origin access identity's information.

See: getCloudFrontOriginAccessIdentity smart constructor.

Request Lenses

gcfoaiId :: Lens' GetCloudFrontOriginAccessIdentity Text

The identity's ID.

Destructuring the Response

getCloudFrontOriginAccessIdentityResponse

Arguments

:: Int
gcfoairsResponseStatus

-> GetCloudFrontOriginAccessIdentityResponse

Creates a value of GetCloudFrontOriginAccessIdentityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcfoairsETag - The current version of the origin access identity's information. For example: E2QWRUHAPOMQZL .

gcfoairsCloudFrontOriginAccessIdentity - The origin access identity's information.

gcfoairsResponseStatus - -- | The response status code.

data GetCloudFrontOriginAccessIdentityResponse

The returned result of the corresponding request.

See: getCloudFrontOriginAccessIdentityResponse smart constructor.

Response Lenses

gcfoairsETag :: Lens' GetCloudFrontOriginAccessIdentityResponse (Maybe Text)

The current version of the origin access identity's information. For example: E2QWRUHAPOMQZL .

gcfoairsCloudFrontOriginAccessIdentity :: Lens' GetCloudFrontOriginAccessIdentityResponse (Maybe CloudFrontOriginAccessIdentity)

The origin access identity's information.

gcfoairsResponseStatus :: Lens' GetCloudFrontOriginAccessIdentityResponse Int

- | The response status code.