Network.AWS.CloudFront.GetDistribution

Description

Get the information about a distribution.

Creating a Request

getDistribution

Arguments

:: Text
gdId

-> GetDistribution

Creates a value of GetDistribution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdId - The distribution's ID.

data GetDistribution

The request to get a distribution's information.

See: getDistribution smart constructor.

Request Lenses

gdId :: Lens' GetDistribution Text

The distribution's ID.

Destructuring the Response

getDistributionResponse

Arguments

:: Int
gdrsResponseStatus

-> GetDistributionResponse

Creates a value of GetDistributionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdrsETag - The current version of the distribution's information. For example: E2QWRUHAPOMQZL .

gdrsDistribution - The distribution's information.

gdrsResponseStatus - -- | The response status code.

data GetDistributionResponse

The returned result of the corresponding request.

See: getDistributionResponse smart constructor.

Response Lenses

gdrsETag :: Lens' GetDistributionResponse (Maybe Text)

The current version of the distribution's information. For example: E2QWRUHAPOMQZL .

gdrsDistribution :: Lens' GetDistributionResponse (Maybe Distribution)

The distribution's information.

gdrsResponseStatus :: Lens' GetDistributionResponse Int

- | The response status code.