Network.AWS.CloudFront.GetDistributionConfig

Description

Get the configuration information about a distribution.

Creating a Request

getDistributionConfig

Arguments

:: Text
gdcId

-> GetDistributionConfig

Creates a value of GetDistributionConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdcId - The distribution's ID.

data GetDistributionConfig

The request to get a distribution configuration.

See: getDistributionConfig smart constructor.

Request Lenses

gdcId :: Lens' GetDistributionConfig Text

The distribution's ID.

Destructuring the Response

getDistributionConfigResponse

Arguments

:: Int
gdcrsResponseStatus

-> GetDistributionConfigResponse

Creates a value of GetDistributionConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdcrsETag - The current version of the configuration. For example: E2QWRUHAPOMQZL .

gdcrsDistributionConfig - The distribution's configuration information.

gdcrsResponseStatus - -- | The response status code.

data GetDistributionConfigResponse

The returned result of the corresponding request.

See: getDistributionConfigResponse smart constructor.

Response Lenses

gdcrsETag :: Lens' GetDistributionConfigResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL .

gdcrsDistributionConfig :: Lens' GetDistributionConfigResponse (Maybe DistributionConfig)

The distribution's configuration information.

gdcrsResponseStatus :: Lens' GetDistributionConfigResponse Int

- | The response status code.