Network.AWS.CloudFront.GetStreamingDistribution

Description

Gets information about a specified RTMP distribution, including the distribution configuration.

Creating a Request

getStreamingDistribution

Arguments

:: Text
gsdId

-> GetStreamingDistribution

Creates a value of GetStreamingDistribution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdId - The streaming distribution's ID.

data GetStreamingDistribution

The request to get a streaming distribution's information.

See: getStreamingDistribution smart constructor.

Request Lenses

gsdId :: Lens' GetStreamingDistribution Text

The streaming distribution's ID.

Destructuring the Response

getStreamingDistributionResponse

Arguments

:: Int
gsdrsResponseStatus

-> GetStreamingDistributionResponse

Creates a value of GetStreamingDistributionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdrsETag - The current version of the streaming distribution's information. For example: E2QWRUHAPOMQZL .

gsdrsStreamingDistribution - The streaming distribution's information.

gsdrsResponseStatus - -- | The response status code.

data GetStreamingDistributionResponse

The returned result of the corresponding request.

See: getStreamingDistributionResponse smart constructor.

Response Lenses

gsdrsETag :: Lens' GetStreamingDistributionResponse (Maybe Text)

The current version of the streaming distribution's information. For example: E2QWRUHAPOMQZL .

gsdrsStreamingDistribution :: Lens' GetStreamingDistributionResponse (Maybe StreamingDistribution)

The streaming distribution's information.

gsdrsResponseStatus :: Lens' GetStreamingDistributionResponse Int

- | The response status code.