Network.AWS.CloudFront.GetStreamingDistributionConfig

Description

Get the configuration information about a streaming distribution.

Creating a Request

getStreamingDistributionConfig

Arguments

:: Text
gsdcId

-> GetStreamingDistributionConfig

Creates a value of GetStreamingDistributionConfig with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdcId - The streaming distribution's ID.

data GetStreamingDistributionConfig

To request to get a streaming distribution configuration.

See: getStreamingDistributionConfig smart constructor.

Request Lenses

gsdcId :: Lens' GetStreamingDistributionConfig Text

The streaming distribution's ID.

Destructuring the Response

getStreamingDistributionConfigResponse

Arguments

:: Int
gsdcrsResponseStatus

-> GetStreamingDistributionConfigResponse

Creates a value of GetStreamingDistributionConfigResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gsdcrsStreamingDistributionConfig - The streaming distribution's configuration information.

gsdcrsETag - The current version of the configuration. For example: E2QWRUHAPOMQZL .

gsdcrsResponseStatus - -- | The response status code.

data GetStreamingDistributionConfigResponse

The returned result of the corresponding request.

See: getStreamingDistributionConfigResponse smart constructor.

Response Lenses

gsdcrsStreamingDistributionConfig :: Lens' GetStreamingDistributionConfigResponse (Maybe StreamingDistributionConfig)

The streaming distribution's configuration information.

gsdcrsETag :: Lens' GetStreamingDistributionConfigResponse (Maybe Text)

The current version of the configuration. For example: E2QWRUHAPOMQZL .

gsdcrsResponseStatus :: Lens' GetStreamingDistributionConfigResponse Int

- | The response status code.