Network.AWS.CloudFront.ListCloudFrontOriginAccessIdentities

Description

List origin access identities.

Request

data ListCloudFrontOriginAccessIdentities

The request to list origin access identities.

Request constructor

mkListCloudFrontOriginAccessIdentities :: ListCloudFrontOriginAccessIdentities

Smart constructor for the minimum required parameters to construct a valid ListCloudFrontOriginAccessIdentities request.

The fields accessible through corresponding lenses are:

Marker :: Maybe Text

MaxItems :: Maybe Text

Request lenses

lcfoaiMarker :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

lcfoaiMaxItems :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

The maximum number of origin access identities you want in the response body.

Response

data ListCloudFrontOriginAccessIdentitiesResponse

The returned result of the corresponding request.

Response constructor

mkListCloudFrontOriginAccessIdentitiesResponse

Arguments

:: CloudFrontOriginAccessIdentityList
lcfoairCloudFrontOriginAccessIdentityList

-> ListCloudFrontOriginAccessIdentitiesResponse

Smart constructor for the minimum required parameters to construct a valid ListCloudFrontOriginAccessIdentitiesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CloudFrontOriginAccessIdentityList :: CloudFrontOriginAccessIdentityList

Response lenses

lcfoairCloudFrontOriginAccessIdentityList :: Lens' ListCloudFrontOriginAccessIdentitiesResponse CloudFrontOriginAccessIdentityList

The CloudFrontOriginAccessIdentityList type.