Network.AWS.CloudFront.ListDistributions

Description

List distributions.

http://docs.aws.amazon.com/AmazonCloudFront/latest/APIReference/ListDistributions.html

Request

data ListDistributions

Request constructor

listDistributions :: ListDistributions

ListDistributions constructor.

The fields accessible through corresponding lenses are:

ldMarker :: Maybe Text

ldMaxItems :: Maybe Text

Request lenses

ldMarker :: Lens' ListDistributions (Maybe Text)

Use this when paginating results to indicate where to begin in your list of distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

ldMaxItems :: Lens' ListDistributions (Maybe Text)

The maximum number of distributions you want in the response body.

Response

data ListDistributionsResponse

Response constructor

listDistributionsResponse

Arguments

:: DistributionList
ldrDistributionList

-> ListDistributionsResponse

ListDistributionsResponse constructor.

The fields accessible through corresponding lenses are:

ldrDistributionList :: DistributionList

Response lenses

ldrDistributionList :: Lens' ListDistributionsResponse DistributionList

The DistributionList type.