Network.AWS.CloudFront.ListInvalidations

Description

List invalidation batches.

http://docs.aws.amazon.com/AmazonCloudFront/latest/APIReference/ListInvalidations.html

Request

data ListInvalidations

Request constructor

listInvalidations

Arguments

:: Text
liDistributionId

-> ListInvalidations

ListInvalidations constructor.

The fields accessible through corresponding lenses are:

liDistributionId :: Text

liMarker :: Maybe Text

liMaxItems :: Maybe Text

Request lenses

liDistributionId :: Lens' ListInvalidations Text

The distribution's id.

liMarker :: Lens' ListInvalidations (Maybe Text)

Use this parameter when paginating results to indicate where to begin in your list of invalidation batches. Because the results are returned in decreasing order from most recent to oldest, the most recent results are on the first page, the second page will contain earlier results, and so on. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response. This value is the same as the ID of the last invalidation batch on that page.

liMaxItems :: Lens' ListInvalidations (Maybe Text)

The maximum number of invalidation batches you want in the response body.

Response

data ListInvalidationsResponse

Response constructor

listInvalidationsResponse

Arguments

:: InvalidationList
lirInvalidationList

-> ListInvalidationsResponse

ListInvalidationsResponse constructor.

The fields accessible through corresponding lenses are:

lirInvalidationList :: InvalidationList

Response lenses

lirInvalidationList :: Lens' ListInvalidationsResponse InvalidationList

Information about invalidation batches.