Network.AWS.CloudFront.ListStreamingDistributions

Description

List streaming distributions.

http://docs.aws.amazon.com/AmazonCloudFront/latest/APIReference/ListStreamingDistributions.html

Request

data ListStreamingDistributions

Request constructor

listStreamingDistributions :: ListStreamingDistributions

ListStreamingDistributions constructor.

The fields accessible through corresponding lenses are:

lsdMarker :: Maybe Text

lsdMaxItems :: Maybe Text

Request lenses

lsdMarker :: Lens' ListStreamingDistributions (Maybe Text)

Use this when paginating results to indicate where to begin in your list of streaming distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

lsdMaxItems :: Lens' ListStreamingDistributions (Maybe Text)

The maximum number of streaming distributions you want in the response body.

Response

data ListStreamingDistributionsResponse

Response constructor

listStreamingDistributionsResponse

Arguments

:: StreamingDistributionList
lsdrStreamingDistributionList

-> ListStreamingDistributionsResponse

ListStreamingDistributionsResponse constructor.

The fields accessible through corresponding lenses are:

lsdrStreamingDistributionList :: StreamingDistributionList

Response lenses

lsdrStreamingDistributionList :: Lens' ListStreamingDistributionsResponse StreamingDistributionList

The StreamingDistributionList type.