Network.AWS.CloudFront.Types

Service

data CloudFront

Version 2014-11-06 of the Amazon CloudFront service.

Error

data RESTError :: *

XML

ns :: Text

CloudFrontOriginAccessIdentityList

data CloudFrontOriginAccessIdentityList

cloudFrontOriginAccessIdentityList

Arguments

:: Text
cfoailMarker

-> Int
cfoailMaxItems

-> Bool
cfoailIsTruncated

-> Int
cfoailQuantity

-> CloudFrontOriginAccessIdentityList

CloudFrontOriginAccessIdentityList constructor.

The fields accessible through corresponding lenses are:

cfoailIsTruncated :: Bool

cfoailItems :: [CloudFrontOriginAccessIdentitySummary]

cfoailMarker :: Text

cfoailMaxItems :: Int

cfoailNextMarker :: Maybe Text

cfoailQuantity :: Int

cfoailIsTruncated :: Lens' CloudFrontOriginAccessIdentityList Bool

A flag that indicates whether more origin access identities remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more items in the list.

cfoailItems :: Lens' CloudFrontOriginAccessIdentityList [CloudFrontOriginAccessIdentitySummary]

A complex type that contains one CloudFrontOriginAccessIdentitySummary element for each origin access identity that was created by the current AWS account.

cfoailMarker :: Lens' CloudFrontOriginAccessIdentityList Text

The value you provided for the Marker request parameter.

cfoailMaxItems :: Lens' CloudFrontOriginAccessIdentityList Int

The value you provided for the MaxItems request parameter.

cfoailNextMarker :: Lens' CloudFrontOriginAccessIdentityList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your origin access identities where they left off.

cfoailQuantity :: Lens' CloudFrontOriginAccessIdentityList Int

The number of CloudFront origin access identities that were created by the current AWS account.

Invalidation

data Invalidation

invalidation

Arguments

:: Text
iId

-> Text
iStatus

-> UTCTime
iCreateTime

-> InvalidationBatch
iInvalidationBatch

-> Invalidation

Invalidation constructor.

The fields accessible through corresponding lenses are:

iCreateTime :: UTCTime

iId :: Text

iInvalidationBatch :: InvalidationBatch

iStatus :: Text

iCreateTime :: Lens' Invalidation UTCTime

The date and time the invalidation request was first made.

iId :: Lens' Invalidation Text

The identifier for the invalidation request. For example: IDFDVBD632BHDS5.

iInvalidationBatch :: Lens' Invalidation InvalidationBatch

The current invalidation information for the batch request.

iStatus :: Lens' Invalidation Text

The status of the invalidation request. When the invalidation batch is finished, the status is Completed.

SSLSupportMethod

data SSLSupportMethod

Constructors

SniOnly
sni-only

Vip
vip

AllowedMethods

data AllowedMethods

allowedMethods

Arguments

:: Int
amQuantity

-> AllowedMethods

AllowedMethods constructor.

The fields accessible through corresponding lenses are:

amCachedMethods :: Maybe CachedMethods

amItems :: [Method]

amQuantity :: Int

amCachedMethods :: Lens' AllowedMethods (Maybe CachedMethods)

amItems :: Lens' AllowedMethods [Method]

A complex type that contains the HTTP methods that you want CloudFront to process and forward to your origin.

amQuantity :: Lens' AllowedMethods Int

The number of HTTP methods that you want CloudFront to forward to your origin. Valid values are 2 (for GET and HEAD requests), 3 (for GET, HEAD and OPTIONS requests) and 7 (for GET, HEAD, OPTIONS, PUT, PATCH, POST, and DELETE requests).

CloudFrontOriginAccessIdentityConfig

data CloudFrontOriginAccessIdentityConfig

cloudFrontOriginAccessIdentityConfig

Arguments

:: Text
cfoaicCallerReference

-> Text
cfoaicComment

-> CloudFrontOriginAccessIdentityConfig

CloudFrontOriginAccessIdentityConfig constructor.

The fields accessible through corresponding lenses are:

cfoaicCallerReference :: Text

cfoaicComment :: Text

cfoaicCallerReference :: Lens' CloudFrontOriginAccessIdentityConfig Text

A unique number that ensures the request can't be replayed. If the CallerReference is new (no matter the content of the CloudFrontOriginAccessIdentityConfig object), a new origin access identity is created. If the CallerReference is a value you already sent in a previous request to create an identity, and the content of the CloudFrontOriginAccessIdentityConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request. If the CallerReference is a value you already sent in a previous request to create an identity but the content of the CloudFrontOriginAccessIdentityConfig is different from the original request, CloudFront returns a CloudFrontOriginAccessIdentityAlreadyExists error.

cfoaicComment :: Lens' CloudFrontOriginAccessIdentityConfig Text

Any comments you want to include about the origin access identity.

Origin

data Origin

origin

Arguments

:: Text
oId

-> Text
oDomainName

-> Origin

Origin constructor.

The fields accessible through corresponding lenses are:

oCustomOriginConfig :: Maybe CustomOriginConfig

oDomainName :: Text

oId :: Text

oOriginPath :: Maybe Text

oS3OriginConfig :: Maybe S3OriginConfig

oCustomOriginConfig :: Lens' Origin (Maybe CustomOriginConfig)

A complex type that contains information about a custom origin. If the origin is an Amazon S3 bucket, use the S3OriginConfig element instead.

oDomainName :: Lens' Origin Text

Amazon S3 origins: The DNS name of the Amazon S3 bucket from which you want CloudFront to get objects for this origin, for example, myawsbucket.s3.amazonaws.com. Custom origins: The DNS domain name for the HTTP server from which you want CloudFront to get objects for this origin, for example, www.example.com.

oId :: Lens' Origin Text

A unique identifier for the origin. The value of Id must be unique within the distribution. You use the value of Id when you create a cache behavior. The Id identifies the origin that CloudFront routes a request to when the request matches the path pattern for that cache behavior.

oOriginPath :: Lens' Origin (Maybe Text)

An optional element that causes CloudFront to request your content from a directory in your Amazon S3 bucket or your custom origin. When you include the OriginPath element, specify the directory name, beginning with a /. CloudFront appends the directory name to the value of DomainName.

oS3OriginConfig :: Lens' Origin (Maybe S3OriginConfig)

A complex type that contains information about the Amazon S3 origin. If the origin is a custom origin, use the CustomOriginConfig element instead.

ViewerProtocolPolicy

data ViewerProtocolPolicy

Constructors

AllowAll
allow-all

HttpsOnly
https-only

RedirectToHttps
redirect-to-https

StreamingDistributionList

data StreamingDistributionList

streamingDistributionList

Arguments

:: Text
sdlMarker

-> Int
sdlMaxItems

-> Bool
sdlIsTruncated

-> Int
sdlQuantity

-> StreamingDistributionList

StreamingDistributionList constructor.

The fields accessible through corresponding lenses are:

sdlIsTruncated :: Bool

sdlItems :: [StreamingDistributionSummary]

sdlMarker :: Text

sdlMaxItems :: Int

sdlNextMarker :: Maybe Text

sdlQuantity :: Int

sdlIsTruncated :: Lens' StreamingDistributionList Bool

A flag that indicates whether more streaming distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

sdlItems :: Lens' StreamingDistributionList [StreamingDistributionSummary]

A complex type that contains one StreamingDistributionSummary element for each distribution that was created by the current AWS account.

sdlMarker :: Lens' StreamingDistributionList Text

The value you provided for the Marker request parameter.

sdlMaxItems :: Lens' StreamingDistributionList Int

The value you provided for the MaxItems request parameter.

sdlNextMarker :: Lens' StreamingDistributionList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your streaming distributions where they left off.

sdlQuantity :: Lens' StreamingDistributionList Int

The number of streaming distributions that were created by the current AWS account.

StreamingDistributionConfig

data StreamingDistributionConfig

streamingDistributionConfig

Arguments

:: Text
sdcCallerReference

-> S3Origin
sdcS3Origin

-> Text
sdcComment

-> TrustedSigners
sdcTrustedSigners

-> Bool
sdcEnabled

-> StreamingDistributionConfig

StreamingDistributionConfig constructor.

The fields accessible through corresponding lenses are:

sdcAliases :: Maybe Aliases

sdcCallerReference :: Text

sdcComment :: Text

sdcEnabled :: Bool

sdcLogging :: Maybe StreamingLoggingConfig

sdcPriceClass :: Maybe PriceClass

sdcS3Origin :: S3Origin

sdcTrustedSigners :: TrustedSigners

sdcAliases :: Lens' StreamingDistributionConfig (Maybe Aliases)

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

sdcCallerReference :: Lens' StreamingDistributionConfig Text

A unique number that ensures the request can't be replayed. If the CallerReference is new (no matter the content of the StreamingDistributionConfig object), a new streaming distribution is created. If the CallerReference is a value you already sent in a previous request to create a streaming distribution, and the content of the StreamingDistributionConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request. If the CallerReference is a value you already sent in a previous request to create a streaming distribution but the content of the StreamingDistributionConfig is different from the original request, CloudFront returns a DistributionAlreadyExists error.

sdcComment :: Lens' StreamingDistributionConfig Text

Any comments you want to include about the streaming distribution.

sdcEnabled :: Lens' StreamingDistributionConfig Bool

Whether the streaming distribution is enabled to accept end user requests for content.

sdcLogging :: Lens' StreamingDistributionConfig (Maybe StreamingLoggingConfig)

A complex type that controls whether access logs are written for the streaming distribution.

sdcPriceClass :: Lens' StreamingDistributionConfig (Maybe PriceClass)

A complex type that contains information about price class for this streaming distribution.

sdcS3Origin :: Lens' StreamingDistributionConfig S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

sdcTrustedSigners :: Lens' StreamingDistributionConfig TrustedSigners

A complex type that specifies the AWS accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items. For more information, go to Using a Signed URL to Serve Private Content in the Amazon CloudFront Developer Guide. If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

Signer

data Signer

signer :: Signer

Signer constructor.

The fields accessible through corresponding lenses are:

sAwsAccountNumber :: Maybe Text

sKeyPairIds :: Maybe KeyPairIds

sAwsAccountNumber :: Lens' Signer (Maybe Text)

Specifies an AWS account that can create signed URLs. Values: self, which indicates that the AWS account that was used to create the distribution can created signed URLs, or an AWS account number. Omit the dashes in the account number.

sKeyPairIds :: Lens' Signer (Maybe KeyPairIds)

A complex type that lists the active CloudFront key pairs, if any, that are associated with AwsAccountNumber.

CookiePreference

data CookiePreference

cookiePreference

Arguments

:: ItemSelection
cpForward

-> CookiePreference

CookiePreference constructor.

The fields accessible through corresponding lenses are:

cpForward :: ItemSelection

cpWhitelistedNames :: Maybe CookieNames

cpForward :: Lens' CookiePreference ItemSelection

Use this element to specify whether you want CloudFront to forward cookies to the origin that is associated with this cache behavior. You can specify all, none or whitelist. If you choose All, CloudFront forwards all cookies regardless of how many your application uses.

cpWhitelistedNames :: Lens' CookiePreference (Maybe CookieNames)

A complex type that specifies the whitelisted cookies, if any, that you want CloudFront to forward to your origin that is associated with this cache behavior.

OriginProtocolPolicy

data OriginProtocolPolicy

Constructors

HttpOnly
http-only

MatchViewer
match-viewer

Distribution

data Distribution

distribution

Arguments

:: Text
dId

-> Text
dStatus

-> UTCTime
dLastModifiedTime

-> Int
dInProgressInvalidationBatches

-> Text
dDomainName

-> ActiveTrustedSigners
dActiveTrustedSigners

-> DistributionConfig
dDistributionConfig

-> Distribution

Distribution constructor.

The fields accessible through corresponding lenses are:

dActiveTrustedSigners :: ActiveTrustedSigners

dDistributionConfig :: DistributionConfig

dDomainName :: Text

dId :: Text

dInProgressInvalidationBatches :: Int

dLastModifiedTime :: UTCTime

dStatus :: Text

dActiveTrustedSigners :: Lens' Distribution ActiveTrustedSigners

CloudFront automatically adds this element to the response only if you've set up the distribution to serve private content with signed URLs. The element lists the key pair IDs that CloudFront is aware of for each trusted signer. The Signer child element lists the AWS account number of the trusted signer (or an empty Self element if the signer is you). The Signer element also includes the IDs of any active key pairs associated with the trusted signer's AWS account. If no KeyPairId element appears for a Signer, that signer can't create working signed URLs.

dDistributionConfig :: Lens' Distribution DistributionConfig

The current configuration information for the distribution.

dDomainName :: Lens' Distribution Text

The domain name corresponding to the distribution. For example: d604721fxaaqy9.cloudfront.net.

dId :: Lens' Distribution Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

dInProgressInvalidationBatches :: Lens' Distribution Int

The number of invalidation batches currently in progress.

dLastModifiedTime :: Lens' Distribution UTCTime

The date and time the distribution was last modified.

dStatus :: Lens' Distribution Text

This response element indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

CloudFrontOriginAccessIdentitySummary

data CloudFrontOriginAccessIdentitySummary

cloudFrontOriginAccessIdentitySummary

Arguments

:: Text
cfoaisId

-> Text
cfoaisS3CanonicalUserId

-> Text
cfoaisComment

-> CloudFrontOriginAccessIdentitySummary

CloudFrontOriginAccessIdentitySummary constructor.

The fields accessible through corresponding lenses are:

cfoaisComment :: Text

cfoaisId :: Text

cfoaisS3CanonicalUserId :: Text

cfoaisComment :: Lens' CloudFrontOriginAccessIdentitySummary Text

The comment for this origin access identity, as originally specified when created.

cfoaisId :: Lens' CloudFrontOriginAccessIdentitySummary Text

The ID for the origin access identity. For example: E74FTE3AJFJ256A.

cfoaisS3CanonicalUserId :: Lens' CloudFrontOriginAccessIdentitySummary Text

The Amazon S3 canonical user ID for the origin access identity, which you use when giving the origin access identity read permission to an object in Amazon S3.

StreamingDistributionSummary

data StreamingDistributionSummary

streamingDistributionSummary

Arguments

:: Text
sdsId

-> Text
sdsStatus

-> UTCTime
sdsLastModifiedTime

-> Text
sdsDomainName

-> S3Origin
sdsS3Origin

-> Aliases
sdsAliases

-> TrustedSigners
sdsTrustedSigners

-> Text
sdsComment

-> PriceClass
sdsPriceClass

-> Bool
sdsEnabled

-> StreamingDistributionSummary

StreamingDistributionSummary constructor.

The fields accessible through corresponding lenses are:

sdsAliases :: Aliases

sdsComment :: Text

sdsDomainName :: Text

sdsEnabled :: Bool

sdsId :: Text

sdsLastModifiedTime :: UTCTime

sdsPriceClass :: PriceClass

sdsS3Origin :: S3Origin

sdsStatus :: Text

sdsTrustedSigners :: TrustedSigners

sdsAliases :: Lens' StreamingDistributionSummary Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this streaming distribution.

sdsComment :: Lens' StreamingDistributionSummary Text

The comment originally specified when this distribution was created.

sdsDomainName :: Lens' StreamingDistributionSummary Text

The domain name corresponding to the distribution. For example: d604721fxaaqy9.cloudfront.net.

sdsEnabled :: Lens' StreamingDistributionSummary Bool

Whether the distribution is enabled to accept end user requests for content.

sdsId :: Lens' StreamingDistributionSummary Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

sdsLastModifiedTime :: Lens' StreamingDistributionSummary UTCTime

The date and time the distribution was last modified.

sdsPriceClass :: Lens' StreamingDistributionSummary PriceClass

sdsS3Origin :: Lens' StreamingDistributionSummary S3Origin

A complex type that contains information about the Amazon S3 bucket from which you want CloudFront to get your media files for distribution.

sdsStatus :: Lens' StreamingDistributionSummary Text

Indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

sdsTrustedSigners :: Lens' StreamingDistributionSummary TrustedSigners

A complex type that specifies the AWS accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items. For more information, go to Using a Signed URL to Serve Private Content in the Amazon CloudFront Developer Guide. If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

CustomOriginConfig

data CustomOriginConfig

customOriginConfig

Arguments

:: Int
cocHTTPPort

-> Int
cocHTTPSPort

-> OriginProtocolPolicy
cocOriginProtocolPolicy

-> CustomOriginConfig

CustomOriginConfig constructor.

The fields accessible through corresponding lenses are:

cocHTTPPort :: Int

cocHTTPSPort :: Int

cocOriginProtocolPolicy :: OriginProtocolPolicy

cocHTTPPort :: Lens' CustomOriginConfig Int

The HTTP port the custom origin listens on.

cocHTTPSPort :: Lens' CustomOriginConfig Int

The HTTPS port the custom origin listens on.

cocOriginProtocolPolicy :: Lens' CustomOriginConfig OriginProtocolPolicy

The origin protocol policy to apply to your origin.

Aliases

data Aliases

aliases

Arguments

:: Int
aQuantity

-> Aliases

Aliases constructor.

The fields accessible through corresponding lenses are:

aItems :: [Text]

aQuantity :: Int

aItems :: Lens' Aliases [Text]

Optional: A complex type that contains CNAME elements, if any, for this distribution. If Quantity is 0, you can omit Items.

aQuantity :: Lens' Aliases Int

The number of CNAMEs, if any, for this distribution.

InvalidationBatch

data InvalidationBatch

invalidationBatch

Arguments

:: Paths
ibPaths

-> Text
ibCallerReference

-> InvalidationBatch

InvalidationBatch constructor.

The fields accessible through corresponding lenses are:

ibCallerReference :: Text

ibPaths :: Paths

ibCallerReference :: Lens' InvalidationBatch Text

A unique name that ensures the request can't be replayed. If the CallerReference is new (no matter the content of the Path object), a new distribution is created. If the CallerReference is a value you already sent in a previous request to create an invalidation batch, and the content of each Path element is identical to the original request, the response includes the same information returned to the original request. If the CallerReference is a value you already sent in a previous request to create a distribution but the content of any Path is different from the original request, CloudFront returns an InvalidationBatchAlreadyExists error.

ibPaths :: Lens' InvalidationBatch Paths

The path of the object to invalidate. The path is relative to the distribution and must begin with a slash (/). You must enclose each invalidation object with the Path element tags. If the path includes non-ASCII characters or unsafe characters as defined in RFC 1783 (http:/www.ietf.orgrfc/rfc1738.txt), URL encode those characters. Do not URL encode any other characters in the path, or CloudFront will not invalidate the old version of the updated object.

InvalidationSummary

data InvalidationSummary

invalidationSummary

Arguments

:: Text
isId

-> UTCTime
isCreateTime

-> Text
isStatus

-> InvalidationSummary

InvalidationSummary constructor.

The fields accessible through corresponding lenses are:

isCreateTime :: UTCTime

isId :: Text

isStatus :: Text

isCreateTime :: Lens' InvalidationSummary UTCTime

isId :: Lens' InvalidationSummary Text

The unique ID for an invalidation request.

isStatus :: Lens' InvalidationSummary Text

The status of an invalidation request.

DistributionConfig

data DistributionConfig

distributionConfig

Arguments

:: Text
dcCallerReference

-> Origins
dcOrigins

-> DefaultCacheBehavior
dcDefaultCacheBehavior

-> Text
dcComment

-> Bool
dcEnabled

-> DistributionConfig

DistributionConfig constructor.

The fields accessible through corresponding lenses are:

dcAliases :: Maybe Aliases

dcCacheBehaviors :: Maybe CacheBehaviors

dcCallerReference :: Text

dcComment :: Text

dcCustomErrorResponses :: Maybe CustomErrorResponses

dcDefaultCacheBehavior :: DefaultCacheBehavior

dcDefaultRootObject :: Maybe Text

dcEnabled :: Bool

dcLogging :: Maybe LoggingConfig

dcOrigins :: Origins

dcPriceClass :: Maybe PriceClass

dcRestrictions :: Maybe Restrictions

dcViewerCertificate :: Maybe ViewerCertificate

dcAliases :: Lens' DistributionConfig (Maybe Aliases)

A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

dcCacheBehaviors :: Lens' DistributionConfig (Maybe CacheBehaviors)

A complex type that contains zero or more CacheBehavior elements.

dcCallerReference :: Lens' DistributionConfig Text

A unique number that ensures the request can't be replayed. If the CallerReference is new (no matter the content of the DistributionConfig object), a new distribution is created. If the CallerReference is a value you already sent in a previous request to create a distribution, and the content of the DistributionConfig is identical to the original request (ignoring white space), the response includes the same information returned to the original request. If the CallerReference is a value you already sent in a previous request to create a distribution but the content of the DistributionConfig is different from the original request, CloudFront returns a DistributionAlreadyExists error.

dcComment :: Lens' DistributionConfig Text

Any comments you want to include about the distribution.

dcCustomErrorResponses :: Lens' DistributionConfig (Maybe CustomErrorResponses)

A complex type that contains zero or more CustomErrorResponse elements.

dcDefaultCacheBehavior :: Lens' DistributionConfig DefaultCacheBehavior

A complex type that describes the default cache behavior if you do not specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements.You must create exactly one default cache behavior.

dcDefaultRootObject :: Lens' DistributionConfig (Maybe Text)

The object that you want CloudFront to return (for example, index.html) when an end user requests the root URL for your distribution (http://www.example.com) instead of an object in your distribution (http:/www.example.comindex.html). Specifying a default root object avoids exposing the contents of your distribution. If you don't want to specify a default root object when you create a distribution, include an empty DefaultRootObject element. To delete the default root object from an existing distribution, update the distribution configuration and include an empty DefaultRootObject element. To replace the default root object, update the distribution configuration and specify the new object.

dcEnabled :: Lens' DistributionConfig Bool

Whether the distribution is enabled to accept end user requests for content.

dcLogging :: Lens' DistributionConfig (Maybe LoggingConfig)

A complex type that controls whether access logs are written for the distribution.

dcOrigins :: Lens' DistributionConfig Origins

A complex type that contains information about origins for this distribution.

dcPriceClass :: Lens' DistributionConfig (Maybe PriceClass)

A complex type that contains information about price class for this distribution.

dcRestrictions :: Lens' DistributionConfig (Maybe Restrictions)

dcViewerCertificate :: Lens' DistributionConfig (Maybe ViewerCertificate)

CacheBehavior

data CacheBehavior

cacheBehavior

Arguments

:: Text
cbPathPattern

-> Text
cbTargetOriginId

-> ForwardedValues
cbForwardedValues

-> TrustedSigners
cbTrustedSigners

-> ViewerProtocolPolicy
cbViewerProtocolPolicy

-> Integer
cbMinTTL

-> CacheBehavior

CacheBehavior constructor.

The fields accessible through corresponding lenses are:

cbAllowedMethods :: Maybe AllowedMethods

cbForwardedValues :: ForwardedValues

cbMinTTL :: Integer

cbPathPattern :: Text

cbSmoothStreaming :: Maybe Bool

cbTargetOriginId :: Text

cbTrustedSigners :: TrustedSigners

cbViewerProtocolPolicy :: ViewerProtocolPolicy

cbAllowedMethods :: Lens' CacheBehavior (Maybe AllowedMethods)

cbForwardedValues :: Lens' CacheBehavior ForwardedValues

A complex type that specifies how CloudFront handles query strings, cookies and headers.

cbMinTTL :: Lens' CacheBehavior Integer

The minimum amount of time that you want objects to stay in CloudFront caches before CloudFront queries your origin to see whether the object has been updated.You can specify a value from 0 to 3,153,600,000 seconds (100 years).

cbPathPattern :: Lens' CacheBehavior Text

The pattern (for example, images/*.jpg) that specifies which requests you want this cache behavior to apply to. When CloudFront receives an end-user request, the requested path is compared with path patterns in the order in which cache behaviors are listed in the distribution. The path pattern for the default cache behavior is * and cannot be changed. If the request for an object does not match the path pattern for any cache behaviors, CloudFront applies the behavior in the default cache behavior.

cbSmoothStreaming :: Lens' CacheBehavior (Maybe Bool)

Indicates whether you want to distribute media files in Microsoft Smooth Streaming format using the origin that is associated with this cache behavior. If so, specify true; if not, specify false.

cbTargetOriginId :: Lens' CacheBehavior Text

The value of ID for the origin that you want CloudFront to route requests to when a request matches the path pattern either for a cache behavior or for the default cache behavior.

cbTrustedSigners :: Lens' CacheBehavior TrustedSigners

A complex type that specifies the AWS accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items. For more information, go to Using a Signed URL to Serve Private Content in the Amazon CloudFront Developer Guide. If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

cbViewerProtocolPolicy :: Lens' CacheBehavior ViewerProtocolPolicy

Use this element to specify the protocol that users can use to access the files in the origin specified by TargetOriginId when a request matches the path pattern in PathPattern. If you want CloudFront to allow end users to use any available protocol, specify allow-all. If you want CloudFront to require HTTPS, specify https. If you want CloudFront to respond to an HTTP request with an HTTP status code of 301 (Moved Permanently) and the HTTPS URL, specify redirect-to-https. The viewer then resubmits the request using the HTTPS URL.

DistributionList

data DistributionList

distributionList

Arguments

:: Text
dlMarker

-> Int
dlMaxItems

-> Bool
dlIsTruncated

-> Int
dlQuantity

-> DistributionList

DistributionList constructor.

The fields accessible through corresponding lenses are:

dlIsTruncated :: Bool

dlItems :: [DistributionSummary]

dlMarker :: Text

dlMaxItems :: Int

dlNextMarker :: Maybe Text

dlQuantity :: Int

dlIsTruncated :: Lens' DistributionList Bool

A flag that indicates whether more distributions remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more distributions in the list.

dlItems :: Lens' DistributionList [DistributionSummary]

A complex type that contains one DistributionSummary element for each distribution that was created by the current AWS account.

dlMarker :: Lens' DistributionList Text

The value you provided for the Marker request parameter.

dlMaxItems :: Lens' DistributionList Int

The value you provided for the MaxItems request parameter.

dlNextMarker :: Lens' DistributionList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your distributions where they left off.

dlQuantity :: Lens' DistributionList Int

The number of distributions that were created by the current AWS account.

KeyPairIds

data KeyPairIds

keyPairIds

Arguments

:: Int
kpiQuantity

-> KeyPairIds

KeyPairIds constructor.

The fields accessible through corresponding lenses are:

kpiItems :: [Text]

kpiQuantity :: Int

kpiItems :: Lens' KeyPairIds [Text]

A complex type that lists the active CloudFront key pairs, if any, that are associated with AwsAccountNumber.

kpiQuantity :: Lens' KeyPairIds Int

The number of active CloudFront key pairs for AwsAccountNumber.

PriceClass

data PriceClass

Constructors

PriceClass100
PriceClass_100

PriceClass200
PriceClass_200

PriceClassAll
PriceClass_All

CustomErrorResponses

data CustomErrorResponses

customErrorResponses

Arguments

:: Int
cerQuantity

-> CustomErrorResponses

CustomErrorResponses constructor.

The fields accessible through corresponding lenses are:

cerItems :: [CustomErrorResponse]

cerQuantity :: Int

cerItems :: Lens' CustomErrorResponses [CustomErrorResponse]

Optional: A complex type that contains custom error responses for this distribution. If Quantity is 0, you can omit Items.

cerQuantity :: Lens' CustomErrorResponses Int

The number of custom error responses for this distribution.

S3OriginConfig

data S3OriginConfig

s3OriginConfig

Arguments

:: Text
socOriginAccessIdentity

-> S3OriginConfig

S3OriginConfig constructor.

The fields accessible through corresponding lenses are:

socOriginAccessIdentity :: Text

socOriginAccessIdentity :: Lens' S3OriginConfig Text

The CloudFront origin access identity to associate with the origin. Use an origin access identity to configure the origin so that end users can only access objects in an Amazon S3 bucket through CloudFront. If you want end users to be able to access objects using either the CloudFront URL or the Amazon S3 URL, specify an empty OriginAccessIdentity element. To delete the origin access identity from an existing distribution, update the distribution configuration and include an empty OriginAccessIdentity element. To replace the origin access identity, update the distribution configuration and specify the new origin access identity. Use the format origin-access-identitycloudfrontId where Id is the value that CloudFront returned in the Id element when you created the origin access identity.

GeoRestriction

data GeoRestriction

geoRestriction

Arguments

:: GeoRestrictionType
grRestrictionType

-> Int
grQuantity

-> GeoRestriction

GeoRestriction constructor.

The fields accessible through corresponding lenses are:

grItems :: [Text]

grQuantity :: Int

grRestrictionType :: GeoRestrictionType

grItems :: Lens' GeoRestriction [Text]

A complex type that contains a Location element for each country in which you want CloudFront either to distribute your content (whitelist) or not distribute your content (blacklist). The Location element is a two-letter, uppercase country code for a country that you want to include in your blacklist or whitelist. Include one Location element for each country. CloudFront and MaxMind both use ISO 3166 country codes. For the current list of countries and the corresponding codes, see ISO 3166-1-alpha-2 code on the International Organization for Standardization website. You can also refer to the country list in the CloudFront console, which includes both country names and codes.

grQuantity :: Lens' GeoRestriction Int

When geo restriction is enabled, this is the number of countries in your whitelist or blacklist. Otherwise, when it is not enabled, Quantity is 0, and you can omit Items.

grRestrictionType :: Lens' GeoRestriction GeoRestrictionType

The method that you want to use to restrict distribution of your content by country: - none: No geo restriction is enabled, meaning access to content is not restricted by client geo location. - blacklist: The Location elements specify the countries in which you do not want CloudFront to distribute your content. - whitelist: The Location elements specify the countries in which you want CloudFront to distribute your content.

S3Origin

data S3Origin

s3Origin

Arguments

:: Text
soDomainName

-> Text
soOriginAccessIdentity

-> S3Origin

S3Origin constructor.

The fields accessible through corresponding lenses are:

soDomainName :: Text

soOriginAccessIdentity :: Text

soDomainName :: Lens' S3Origin Text

The DNS name of the S3 origin.

soOriginAccessIdentity :: Lens' S3Origin Text

Your S3 origin's origin access identity.

Headers

data Headers

headers

Arguments

:: Int
hQuantity

-> Headers

Headers constructor.

The fields accessible through corresponding lenses are:

hItems :: [Text]

hQuantity :: Int

hItems :: Lens' Headers [Text]

Optional: A complex type that contains a Name element for each header that you want CloudFront to forward to the origin and to vary on for this cache behavior. If Quantity is 0, omit Items.

hQuantity :: Lens' Headers Int

The number of different headers that you want CloudFront to forward to the origin and to vary on for this cache behavior. The maximum number of headers that you can specify by name is 10. If you want CloudFront to forward all headers to the origin and vary on all of them, specify 1 for Quantity and * for Name. If you don't want CloudFront to forward any additional headers to the origin or to vary on any headers, specify 0 for Quantity and omit Items.

CachedMethods

data CachedMethods

cachedMethods

Arguments

:: Int
cmQuantity

-> CachedMethods

CachedMethods constructor.

The fields accessible through corresponding lenses are:

cmItems :: [Method]

cmQuantity :: Int

cmItems :: Lens' CachedMethods [Method]

A complex type that contains the HTTP methods that you want CloudFront to cache responses to.

cmQuantity :: Lens' CachedMethods Int

The number of HTTP methods for which you want CloudFront to cache responses. Valid values are 2 (for caching responses to GET and HEAD requests) and 3 (for caching responses to GET, HEAD, and OPTIONS requests).

ViewerCertificate

data ViewerCertificate

viewerCertificate :: ViewerCertificate

ViewerCertificate constructor.

The fields accessible through corresponding lenses are:

vcCloudFrontDefaultCertificate :: Maybe Bool

vcIAMCertificateId :: Maybe Text

vcMinimumProtocolVersion :: Maybe MinimumProtocolVersion

vcSSLSupportMethod :: Maybe SSLSupportMethod

vcCloudFrontDefaultCertificate :: Lens' ViewerCertificate (Maybe Bool)

If you want viewers to use HTTPS to request your objects and you're using the CloudFront domain name of your distribution in your object URLs (for example, https://d111111abcdef8.cloudfront.net/logo.jpg), set to true. Omit this value if you are setting an IAMCertificateId.

vcIAMCertificateId :: Lens' ViewerCertificate (Maybe Text)

If you want viewers to use HTTPS to request your objects and you're using an alternate domain name in your object URLs (for example, https://example.com/logo.jpg), specify the IAM certificate identifier of the custom viewer certificate for this distribution. Specify either this value or CloudFrontDefaultCertificate.

vcMinimumProtocolVersion :: Lens' ViewerCertificate (Maybe MinimumProtocolVersion)

Specify the minimum version of the SSL protocol that you want CloudFront to use, SSLv3 or TLSv1, for HTTPS connections. CloudFront will serve your objects only to browsers or devices that support at least the SSL version that you specify. The TLSv1 protocol is more secure, so we recommend that you specify SSLv3 only if your users are using browsers or devices that don't support TLSv1. If you're using a custom certificate (if you specify a value for IAMCertificateId) and if you're using dedicated IP (if you specify vip for SSLSupportMethod), you can choose SSLv3 or TLSv1 as the MinimumProtocolVersion. If you're using a custom certificate (if you specify a value for IAMCertificateId) and if you're using SNI (if you specify sni-only for SSLSupportMethod), you must specify TLSv1 for MinimumProtocolVersion.

vcSSLSupportMethod :: Lens' ViewerCertificate (Maybe SSLSupportMethod)

If you specify a value for IAMCertificateId, you must also specify how you want CloudFront to serve HTTPS requests. Valid values are vip and sni-only. If you specify vip, CloudFront uses dedicated IP addresses for your content and can respond to HTTPS requests from any viewer. However, you must request permission to use this feature, and you incur additional monthly charges. If you specify sni-only, CloudFront can only respond to HTTPS requests from viewers that support Server Name Indication (SNI). All modern browsers support SNI, but some browsers still in use don't support SNI. Do not specify a value for SSLSupportMethod if you specified true for CloudFrontDefaultCertificate.

Restrictions

data Restrictions

restrictions

Arguments

:: GeoRestriction
rGeoRestriction

-> Restrictions

Restrictions constructor.

The fields accessible through corresponding lenses are:

rGeoRestriction :: GeoRestriction

rGeoRestriction :: Lens' Restrictions GeoRestriction

Origins

data Origins

origins

Arguments

:: Int
oQuantity

-> NonEmpty Origin
oItems

-> Origins

Origins constructor.

The fields accessible through corresponding lenses are:

oItems :: NonEmpty Origin

oQuantity :: Int

oItems :: Lens' Origins (NonEmpty Origin)

A complex type that contains origins for this distribution.

oQuantity :: Lens' Origins Int

The number of origins for this distribution.

Method

data Method

Constructors

Delete'
DELETE

Get
GET

Head'
HEAD

Options
OPTIONS

Patch
PATCH

Post
POST

Put
PUT

MinimumProtocolVersion

data MinimumProtocolVersion

Constructors

SSLv3
SSLv3

TLSv1
TLSv1

ForwardedValues

data ForwardedValues

forwardedValues

Arguments

:: Bool
fvQueryString

-> CookiePreference
fvCookies

-> ForwardedValues

ForwardedValues constructor.

The fields accessible through corresponding lenses are:

fvCookies :: CookiePreference

fvHeaders :: Maybe Headers

fvQueryString :: Bool

fvCookies :: Lens' ForwardedValues CookiePreference

A complex type that specifies how CloudFront handles cookies.

fvHeaders :: Lens' ForwardedValues (Maybe Headers)

A complex type that specifies the Headers, if any, that you want CloudFront to vary upon for this cache behavior.

fvQueryString :: Lens' ForwardedValues Bool

Indicates whether you want CloudFront to forward query strings to the origin that is associated with this cache behavior. If so, specify true; if not, specify false.

TrustedSigners

data TrustedSigners

trustedSigners

Arguments

:: Bool
tsEnabled

-> Int
tsQuantity

-> TrustedSigners

TrustedSigners constructor.

The fields accessible through corresponding lenses are:

tsEnabled :: Bool

tsItems :: [Text]

tsQuantity :: Int

tsEnabled :: Lens' TrustedSigners Bool

Specifies whether you want to require end users to use signed URLs to access the files specified by PathPattern and TargetOriginId.

tsItems :: Lens' TrustedSigners [Text]

Optional: A complex type that contains trusted signers for this cache behavior. If Quantity is 0, you can omit Items.

tsQuantity :: Lens' TrustedSigners Int

The number of trusted signers for this cache behavior.

ItemSelection

data ItemSelection

Constructors

All
all

None
none

Whitelist
whitelist

StreamingLoggingConfig

data StreamingLoggingConfig

streamingLoggingConfig

Arguments

:: Bool
slcEnabled

-> Text
slcBucket

-> Text
slcPrefix

-> StreamingLoggingConfig

StreamingLoggingConfig constructor.

The fields accessible through corresponding lenses are:

slcBucket :: Text

slcEnabled :: Bool

slcPrefix :: Text

slcBucket :: Lens' StreamingLoggingConfig Text

The Amazon S3 bucket to store the access logs in, for example, myawslogbucket.s3.amazonaws.com.

slcEnabled :: Lens' StreamingLoggingConfig Bool

Specifies whether you want CloudFront to save access logs to an Amazon S3 bucket. If you do not want to enable logging when you create a streaming distribution or if you want to disable logging for an existing streaming distribution, specify false for Enabled, and specify empty Bucket and Prefix elements. If you specify false for Enabled but you specify values for Bucket and Prefix, the values are automatically deleted.

slcPrefix :: Lens' StreamingLoggingConfig Text

An optional string that you want CloudFront to prefix to the access log filenames for this streaming distribution, for example, myprefix/. If you want to enable logging, but you do not want to specify a prefix, you still must include an empty Prefix element in the Logging element.

CookieNames

data CookieNames

cookieNames

Arguments

:: Int
cnQuantity

-> CookieNames

CookieNames constructor.

The fields accessible through corresponding lenses are:

cnItems :: [Text]

cnQuantity :: Int

cnItems :: Lens' CookieNames [Text]

Optional: A complex type that contains whitelisted cookies for this cache behavior. If Quantity is 0, you can omit Items.

cnQuantity :: Lens' CookieNames Int

The number of whitelisted cookies for this cache behavior.

CustomErrorResponse

data CustomErrorResponse

customErrorResponse

Arguments

:: Int
cerErrorCode

-> CustomErrorResponse

CustomErrorResponse constructor.

The fields accessible through corresponding lenses are:

cerErrorCachingMinTTL :: Maybe Integer

cerErrorCode :: Int

cerResponseCode :: Maybe Text

cerResponsePagePath :: Maybe Text

cerErrorCachingMinTTL :: Lens' CustomErrorResponse (Maybe Integer)

The minimum amount of time you want HTTP error codes to stay in CloudFront caches before CloudFront queries your origin to see whether the object has been updated. You can specify a value from 0 to 31,536,000.

cerErrorCode :: Lens' CustomErrorResponse Int

The 4xx or 5xx HTTP status code that you want to customize. For a list of HTTP status codes that you can customize, see CloudFront documentation.

cerResponseCode :: Lens' CustomErrorResponse (Maybe Text)

The HTTP status code that you want CloudFront to return with the custom error page to the viewer. For a list of HTTP status codes that you can replace, see CloudFront Documentation.

cerResponsePagePath :: Lens' CustomErrorResponse (Maybe Text)

The path of the custom error page (for example, /custom_404.html). The path is relative to the distribution and must begin with a slash (/). If the path includes any non-ASCII characters or unsafe characters as defined in RFC 1783 (http:/www.ietf.orgrfc/rfc1738.txt), URL encode those characters. Do not URL encode any other characters in the path, or CloudFront will not return the custom error page to the viewer.

CacheBehaviors

data CacheBehaviors

cacheBehaviors

Arguments

:: Int
cbQuantity

-> CacheBehaviors

CacheBehaviors constructor.

The fields accessible through corresponding lenses are:

cbItems :: [CacheBehavior]

cbQuantity :: Int

cbItems :: Lens' CacheBehaviors [CacheBehavior]

Optional: A complex type that contains cache behaviors for this distribution. If Quantity is 0, you can omit Items.

cbQuantity :: Lens' CacheBehaviors Int

The number of cache behaviors for this distribution.

DefaultCacheBehavior

data DefaultCacheBehavior

defaultCacheBehavior

Arguments

:: Text
dcbTargetOriginId

-> ForwardedValues
dcbForwardedValues

-> TrustedSigners
dcbTrustedSigners

-> ViewerProtocolPolicy
dcbViewerProtocolPolicy

-> Integer
dcbMinTTL

-> DefaultCacheBehavior

DefaultCacheBehavior constructor.

The fields accessible through corresponding lenses are:

dcbAllowedMethods :: Maybe AllowedMethods

dcbForwardedValues :: ForwardedValues

dcbMinTTL :: Integer

dcbSmoothStreaming :: Maybe Bool

dcbTargetOriginId :: Text

dcbTrustedSigners :: TrustedSigners

dcbViewerProtocolPolicy :: ViewerProtocolPolicy

dcbAllowedMethods :: Lens' DefaultCacheBehavior (Maybe AllowedMethods)

dcbForwardedValues :: Lens' DefaultCacheBehavior ForwardedValues

A complex type that specifies how CloudFront handles query strings, cookies and headers.

dcbMinTTL :: Lens' DefaultCacheBehavior Integer

The minimum amount of time that you want objects to stay in CloudFront caches before CloudFront queries your origin to see whether the object has been updated.You can specify a value from 0 to 3,153,600,000 seconds (100 years).

dcbSmoothStreaming :: Lens' DefaultCacheBehavior (Maybe Bool)

Indicates whether you want to distribute media files in Microsoft Smooth Streaming format using the origin that is associated with this cache behavior. If so, specify true; if not, specify false.

dcbTargetOriginId :: Lens' DefaultCacheBehavior Text

The value of ID for the origin that you want CloudFront to route requests to when a request matches the path pattern either for a cache behavior or for the default cache behavior.

dcbTrustedSigners :: Lens' DefaultCacheBehavior TrustedSigners

A complex type that specifies the AWS accounts, if any, that you want to allow to create signed URLs for private content. If you want to require signed URLs in requests for objects in the target origin that match the PathPattern for this cache behavior, specify true for Enabled, and specify the applicable values for Quantity and Items. For more information, go to Using a Signed URL to Serve Private Content in the Amazon CloudFront Developer Guide. If you don't want to require signed URLs in requests for objects that match PathPattern, specify false for Enabled and 0 for Quantity. Omit Items. To add, change, or remove one or more trusted signers, change Enabled to true (if it's currently false), change Quantity as applicable, and specify all of the trusted signers that you want to include in the updated distribution.

dcbViewerProtocolPolicy :: Lens' DefaultCacheBehavior ViewerProtocolPolicy

Use this element to specify the protocol that users can use to access the files in the origin specified by TargetOriginId when a request matches the path pattern in PathPattern. If you want CloudFront to allow end users to use any available protocol, specify allow-all. If you want CloudFront to require HTTPS, specify https. If you want CloudFront to respond to an HTTP request with an HTTP status code of 301 (Moved Permanently) and the HTTPS URL, specify redirect-to-https. The viewer then resubmits the request using the HTTPS URL.

InvalidationList

data InvalidationList

invalidationList

Arguments

:: Text
ilMarker

-> Int
ilMaxItems

-> Bool
ilIsTruncated

-> Int
ilQuantity

-> InvalidationList

InvalidationList constructor.

The fields accessible through corresponding lenses are:

ilIsTruncated :: Bool

ilItems :: [InvalidationSummary]

ilMarker :: Text

ilMaxItems :: Int

ilNextMarker :: Maybe Text

ilQuantity :: Int

ilIsTruncated :: Lens' InvalidationList Bool

A flag that indicates whether more invalidation batch requests remain to be listed. If your results were truncated, you can make a follow-up pagination request using the Marker request parameter to retrieve more invalidation batches in the list.

ilItems :: Lens' InvalidationList [InvalidationSummary]

A complex type that contains one InvalidationSummary element for each invalidation batch that was created by the current AWS account.

ilMarker :: Lens' InvalidationList Text

The value you provided for the Marker request parameter.

ilMaxItems :: Lens' InvalidationList Int

The value you provided for the MaxItems request parameter.

ilNextMarker :: Lens' InvalidationList (Maybe Text)

If IsTruncated is true, this element is present and contains the value you can use for the Marker request parameter to continue listing your invalidation batches where they left off.

ilQuantity :: Lens' InvalidationList Int

The number of invalidation batches that were created by the current AWS account.

StreamingDistribution

data StreamingDistribution

streamingDistribution

Arguments

:: Text
sdId

-> Text
sdStatus

-> Text
sdDomainName

-> ActiveTrustedSigners
sdActiveTrustedSigners

-> StreamingDistributionConfig
sdStreamingDistributionConfig

-> StreamingDistribution

StreamingDistribution constructor.

The fields accessible through corresponding lenses are:

sdActiveTrustedSigners :: ActiveTrustedSigners

sdDomainName :: Text

sdId :: Text

sdLastModifiedTime :: Maybe UTCTime

sdStatus :: Text

sdStreamingDistributionConfig :: StreamingDistributionConfig

sdActiveTrustedSigners :: Lens' StreamingDistribution ActiveTrustedSigners

CloudFront automatically adds this element to the response only if you've set up the distribution to serve private content with signed URLs. The element lists the key pair IDs that CloudFront is aware of for each trusted signer. The Signer child element lists the AWS account number of the trusted signer (or an empty Self element if the signer is you). The Signer element also includes the IDs of any active key pairs associated with the trusted signer's AWS account. If no KeyPairId element appears for a Signer, that signer can't create working signed URLs.

sdDomainName :: Lens' StreamingDistribution Text

The domain name corresponding to the streaming distribution. For example: s5c39gqb8ow64r.cloudfront.net.

sdId :: Lens' StreamingDistribution Text

The identifier for the streaming distribution. For example: EGTXBD79H29TRA8.

sdLastModifiedTime :: Lens' StreamingDistribution (Maybe UTCTime)

The date and time the distribution was last modified.

sdStatus :: Lens' StreamingDistribution Text

The current status of the streaming distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

sdStreamingDistributionConfig :: Lens' StreamingDistribution StreamingDistributionConfig

The current configuration information for the streaming distribution.

Paths

data Paths

paths

Arguments

:: Int
pQuantity

-> Paths

Paths constructor.

The fields accessible through corresponding lenses are:

pItems :: [Text]

pQuantity :: Int

pItems :: Lens' Paths [Text]

A complex type that contains a list of the objects that you want to invalidate.

pQuantity :: Lens' Paths Int

The number of objects that you want to invalidate.

CloudFrontOriginAccessIdentity

data CloudFrontOriginAccessIdentity

cloudFrontOriginAccessIdentity

Arguments

:: Text
cfoaiId

-> Text
cfoaiS3CanonicalUserId

-> CloudFrontOriginAccessIdentity

CloudFrontOriginAccessIdentity constructor.

The fields accessible through corresponding lenses are:

cfoaiCloudFrontOriginAccessIdentityConfig :: Maybe CloudFrontOriginAccessIdentityConfig

cfoaiId :: Text

cfoaiS3CanonicalUserId :: Text

cfoaiCloudFrontOriginAccessIdentityConfig :: Lens' CloudFrontOriginAccessIdentity (Maybe CloudFrontOriginAccessIdentityConfig)

The current configuration information for the identity.

cfoaiId :: Lens' CloudFrontOriginAccessIdentity Text

The ID for the origin access identity. For example: E74FTE3AJFJ256A.

cfoaiS3CanonicalUserId :: Lens' CloudFrontOriginAccessIdentity Text

The Amazon S3 canonical user ID for the origin access identity, which you use when giving the origin access identity read permission to an object in Amazon S3.

ActiveTrustedSigners

data ActiveTrustedSigners

activeTrustedSigners

Arguments

:: Bool
atsEnabled

-> Int
atsQuantity

-> ActiveTrustedSigners

ActiveTrustedSigners constructor.

The fields accessible through corresponding lenses are:

atsEnabled :: Bool

atsItems :: [Signer]

atsQuantity :: Int

atsEnabled :: Lens' ActiveTrustedSigners Bool

Each active trusted signer.

atsItems :: Lens' ActiveTrustedSigners [Signer]

A complex type that contains one Signer complex type for each unique trusted signer that is specified in the TrustedSigners complex type, including trusted signers in the default cache behavior and in all of the other cache behaviors.

atsQuantity :: Lens' ActiveTrustedSigners Int

The number of unique trusted signers included in all cache behaviors. For example, if three cache behaviors all list the same three AWS accounts, the value of Quantity for ActiveTrustedSigners will be 3.

DistributionSummary

data DistributionSummary

distributionSummary

Arguments

:: Text
dsId

-> Text
dsStatus

-> UTCTime
dsLastModifiedTime

-> Text
dsDomainName

-> Aliases
dsAliases

-> Origins
dsOrigins

-> DefaultCacheBehavior
dsDefaultCacheBehavior

-> CacheBehaviors
dsCacheBehaviors

-> CustomErrorResponses
dsCustomErrorResponses

-> Text
dsComment

-> PriceClass
dsPriceClass

-> Bool
dsEnabled

-> ViewerCertificate
dsViewerCertificate

-> Restrictions
dsRestrictions

-> DistributionSummary

DistributionSummary constructor.

The fields accessible through corresponding lenses are:

dsAliases :: Aliases

dsCacheBehaviors :: CacheBehaviors

dsComment :: Text

dsCustomErrorResponses :: CustomErrorResponses

dsDefaultCacheBehavior :: DefaultCacheBehavior

dsDomainName :: Text

dsEnabled :: Bool

dsId :: Text

dsLastModifiedTime :: UTCTime

dsOrigins :: Origins

dsPriceClass :: PriceClass

dsRestrictions :: Restrictions

dsStatus :: Text

dsViewerCertificate :: ViewerCertificate

dsAliases :: Lens' DistributionSummary Aliases

A complex type that contains information about CNAMEs (alternate domain names), if any, for this distribution.

dsCacheBehaviors :: Lens' DistributionSummary CacheBehaviors

A complex type that contains zero or more CacheBehavior elements.

dsComment :: Lens' DistributionSummary Text

The comment originally specified when this distribution was created.

dsCustomErrorResponses :: Lens' DistributionSummary CustomErrorResponses

A complex type that contains zero or more CustomErrorResponses elements.

dsDefaultCacheBehavior :: Lens' DistributionSummary DefaultCacheBehavior

A complex type that describes the default cache behavior if you do not specify a CacheBehavior element or if files don't match any of the values of PathPattern in CacheBehavior elements.You must create exactly one default cache behavior.

dsDomainName :: Lens' DistributionSummary Text

The domain name corresponding to the distribution. For example: d604721fxaaqy9.cloudfront.net.

dsEnabled :: Lens' DistributionSummary Bool

Whether the distribution is enabled to accept end user requests for content.

dsId :: Lens' DistributionSummary Text

The identifier for the distribution. For example: EDFDVBD632BHDS5.

dsLastModifiedTime :: Lens' DistributionSummary UTCTime

The date and time the distribution was last modified.

dsOrigins :: Lens' DistributionSummary Origins

A complex type that contains information about origins for this distribution.

dsPriceClass :: Lens' DistributionSummary PriceClass

dsRestrictions :: Lens' DistributionSummary Restrictions

dsStatus :: Lens' DistributionSummary Text

This response element indicates the current status of the distribution. When the status is Deployed, the distribution's information is fully propagated throughout the Amazon CloudFront system.

dsViewerCertificate :: Lens' DistributionSummary ViewerCertificate

GeoRestrictionType

data GeoRestrictionType

Constructors

GRTBlacklist
blacklist

GRTNone
none

GRTWhitelist
whitelist

LoggingConfig

data LoggingConfig

loggingConfig

Arguments

:: Bool
lcEnabled

-> Bool
lcIncludeCookies

-> Text
lcBucket

-> Text
lcPrefix

-> LoggingConfig

LoggingConfig constructor.

The fields accessible through corresponding lenses are:

lcBucket :: Text

lcEnabled :: Bool

lcIncludeCookies :: Bool

lcPrefix :: Text

lcBucket :: Lens' LoggingConfig Text

The Amazon S3 bucket to store the access logs in, for example, myawslogbucket.s3.amazonaws.com.

lcEnabled :: Lens' LoggingConfig Bool

Specifies whether you want CloudFront to save access logs to an Amazon S3 bucket. If you do not want to enable logging when you create a distribution or if you want to disable logging for an existing distribution, specify false for Enabled, and specify empty Bucket and Prefix elements. If you specify false for Enabled but you specify values for Bucket, prefix and IncludeCookies, the values are automatically deleted.

lcIncludeCookies :: Lens' LoggingConfig Bool

Specifies whether you want CloudFront to include cookies in access logs, specify true for IncludeCookies. If you choose to include cookies in logs, CloudFront logs all cookies regardless of how you configure the cache behaviors for this distribution. If you do not want to include cookies when you create a distribution or if you want to disable include cookies for an existing distribution, specify false for IncludeCookies.

lcPrefix :: Lens' LoggingConfig Text

An optional string that you want CloudFront to prefix to the access log filenames for this distribution, for example, myprefix/. If you want to enable logging, but you do not want to specify a prefix, you still must include an empty Prefix element in the Logging element.