Network.AWS.CloudFront.V2014_05_31

Description

Amazon CloudFront is a web service that speeds up distribution of your static and dynamic web content, for example, .html, .css, .php, image, and media files, to end users. CloudFront delivers your content through a worldwide network of edge locations. When an end user requests content that you're serving with CloudFront, the user is routed to the edge location that provides the lowest latency, so content is delivered with the best possible performance. If the content is already in that edge location, CloudFront delivers it immediately. If the content is not currently in that edge location, CloudFront retrieves it from an Amazon S3 bucket or an HTTP server (for example, a web server) that you have identified as the source for the definitive version of your content.

Documentation

module Network.AWS.CloudFront.V2014_05_31.CreateCloudFrontOriginAccessIdentity

module Network.AWS.CloudFront.V2014_05_31.CreateDistribution

module Network.AWS.CloudFront.V2014_05_31.CreateInvalidation

module Network.AWS.CloudFront.V2014_05_31.CreateStreamingDistribution

module Network.AWS.CloudFront.V2014_05_31.DeleteCloudFrontOriginAccessIdentity

module Network.AWS.CloudFront.V2014_05_31.DeleteDistribution

module Network.AWS.CloudFront.V2014_05_31.DeleteStreamingDistribution

module Network.AWS.CloudFront.V2014_05_31.GetCloudFrontOriginAccessIdentity

module Network.AWS.CloudFront.V2014_05_31.GetCloudFrontOriginAccessIdentityConfig

module Network.AWS.CloudFront.V2014_05_31.GetDistribution

module Network.AWS.CloudFront.V2014_05_31.GetDistributionConfig

module Network.AWS.CloudFront.V2014_05_31.GetInvalidation

module Network.AWS.CloudFront.V2014_05_31.GetStreamingDistribution

module Network.AWS.CloudFront.V2014_05_31.GetStreamingDistributionConfig

module Network.AWS.CloudFront.V2014_05_31.ListCloudFrontOriginAccessIdentities

module Network.AWS.CloudFront.V2014_05_31.ListDistributions

module Network.AWS.CloudFront.V2014_05_31.ListInvalidations

module Network.AWS.CloudFront.V2014_05_31.ListStreamingDistributions

module Network.AWS.CloudFront.V2014_05_31.Types

module Network.AWS.CloudFront.V2014_05_31.UpdateCloudFrontOriginAccessIdentity

module Network.AWS.CloudFront.V2014_05_31.UpdateDistribution

module Network.AWS.CloudFront.V2014_05_31.UpdateStreamingDistribution