Network.AWS.CloudFront.V2014_05_31.GetStreamingDistribution

Description

Get the information about a streaming distribution.

Request

data GetStreamingDistribution

The request to get a streaming distribution's information.

Request constructor

mkGetStreamingDistribution

Arguments

:: Text
gsdId

-> GetStreamingDistribution

Smart constructor for the minimum required parameters to construct a valid GetStreamingDistribution request.

The fields accessible through corresponding lenses are:

Id :: Text

Request lenses

gsdId :: Lens' GetStreamingDistribution Text

The streaming distribution's id.

Response

data GetStreamingDistributionResponse

The returned result of the corresponding request.

Response constructor

mkGetStreamingDistributionResponse :: GetStreamingDistributionResponse

Smart constructor for the minimum required parameters to construct a valid GetStreamingDistributionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

StreamingDistribution :: Maybe StreamingDistribution

ETag :: Maybe Text

Response lenses

gsdrStreamingDistribution :: Lens' GetStreamingDistributionResponse (Maybe StreamingDistribution)

The streaming distribution's information.

gsdrETag :: Lens' GetStreamingDistributionResponse (Maybe Text)

The current version of the streaming distribution's information. For example: E2QWRUHAPOMQZL.