Network.AWS.CloudFront.V2014_05_31.ListCloudFrontOriginAccessIdentities

Description

List origin access identities.

Documentation

listCloudFrontOriginAccessIdentities :: ListCloudFrontOriginAccessIdentities

Minimum specification for a ListCloudFrontOriginAccessIdentities request.

data ListCloudFrontOriginAccessIdentities

Constructors

ListCloudFrontOriginAccessIdentities

Fields

_lcfoairMarker :: Maybe Text

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

_lcfoairMaxItems :: Maybe Text

The maximum number of origin access identities you want in the response body.

lcfoairMaxItems :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

lcfoairMarker :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

data ListCloudFrontOriginAccessIdentitiesResponse

Constructors

ListCloudFrontOriginAccessIdentitiesResponse

Fields

_lcfoaisCloudFrontOriginAccessIdentityList :: CloudFrontOriginAccessIdentityList

The CloudFrontOriginAccessIdentityList type.

lcfoaisCloudFrontOriginAccessIdentityList :: Iso' ListCloudFrontOriginAccessIdentitiesResponse CloudFrontOriginAccessIdentityList