Network.AWS.CloudFront.V2014_05_31.ListInvalidations

Description

List invalidation batches.

Documentation

listInvalidations

Arguments

:: Text
_lirDistributionId

-> ListInvalidations

Minimum specification for a ListInvalidations request.

data ListInvalidations

Constructors

ListInvalidations

Fields

_lirDistributionId :: Text

The distribution's id.

_lirMarker :: Maybe Text

Use this parameter when paginating results to indicate where to begin in your list of invalidation batches. Because the results are returned in decreasing order from most recent to oldest, the most recent results are on the first page, the second page will contain earlier results, and so on. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response. This value is the same as the ID of the last invalidation batch on that page.

_lirMaxItems :: Maybe Text

The maximum number of invalidation batches you want in the response body.

lirMaxItems :: Lens' ListInvalidations (Maybe Text)

lirMarker :: Lens' ListInvalidations (Maybe Text)

lirDistributionId :: Lens' ListInvalidations Text

data ListInvalidationsResponse

Constructors

ListInvalidationsResponse

Fields

_lisInvalidationList :: InvalidationList

Information about invalidation batches.

lisInvalidationList :: Iso' ListInvalidationsResponse InvalidationList