Network.AWS.CloudFront.V2014_05_31.ListStreamingDistributions

Description

List streaming distributions.

Request

data ListStreamingDistributions

The request to list your streaming distributions.

Request constructor

mkListStreamingDistributions :: ListStreamingDistributions

Smart constructor for the minimum required parameters to construct a valid ListStreamingDistributions request.

Request lenses

lsdMarker :: Lens' ListStreamingDistributions (Maybe Text)

Use this when paginating results to indicate where to begin in your list of streaming distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

lsdMaxItems :: Lens' ListStreamingDistributions (Maybe Text)

The maximum number of streaming distributions you want in the response body.

Response

data ListStreamingDistributionsResponse

The returned result of the corresponding request.

Response lenses

lsdrsStreamingDistributionList :: Lens' ListStreamingDistributionsResponse StreamingDistributionList

The StreamingDistributionList type.