Network.AWS.CloudFront.V2014_05_31.Monadic

Description

Amazon CloudFront is a web service that speeds up distribution of your static and dynamic web content, for example, .html, .css, .php, image, and media files, to end users. CloudFront delivers your content through a worldwide network of edge locations. When an end user requests content that you're serving with CloudFront, the user is routed to the edge location that provides the lowest latency, so content is delivered with the best possible performance. If the content is already in that edge location, CloudFront delivers it immediately. If the content is not currently in that edge location, CloudFront retrieves it from an Amazon S3 bucket or an HTTP server (for example, a web server) that you have identified as the source for the definitive version of your content.

The State operator variants from Control.Lens.Setter such as .= can be used to modify any additional request parameters before sending.

CreateCloudFrontOriginAccessIdentity

Create a new origin access identity.

See: CreateCloudFrontOriginAccessIdentity

createCloudFrontOriginAccessIdentity

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> CloudFrontOriginAccessIdentityConfig
ccfoaiCloudFrontOriginAccessIdentityConfig

-> State CreateCloudFrontOriginAccessIdentity a

-> m CreateCloudFrontOriginAccessIdentityResponse

createCloudFrontOriginAccessIdentityCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> CloudFrontOriginAccessIdentityConfig
ccfoaiCloudFrontOriginAccessIdentityConfig

-> State CreateCloudFrontOriginAccessIdentity a

-> m (Either ServiceEr CreateCloudFrontOriginAccessIdentityResponse)

CreateDistribution

Create a new distribution.

See: CreateDistribution

createDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> DistributionConfig
cdDistributionConfig

-> State CreateDistribution a

-> m CreateDistributionResponse

createDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> DistributionConfig
cdDistributionConfig

-> State CreateDistribution a

-> m (Either ServiceEr CreateDistributionResponse)

CreateInvalidation

Create a new invalidation.

See: CreateInvalidation

createInvalidation

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ciDistributionId

-> InvalidationBatch
ciInvalidationBatch

-> State CreateInvalidation a

-> m CreateInvalidationResponse

createInvalidationCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ciDistributionId

-> InvalidationBatch
ciInvalidationBatch

-> State CreateInvalidation a

-> m (Either ServiceEr CreateInvalidationResponse)

CreateStreamingDistribution

Create a new streaming distribution.

See: CreateStreamingDistribution

createStreamingDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> StreamingDistributionConfig
csdStreamingDistributionConfig

-> State CreateStreamingDistribution a

-> m CreateStreamingDistributionResponse

createStreamingDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> StreamingDistributionConfig
csdStreamingDistributionConfig

-> State CreateStreamingDistribution a

-> m (Either ServiceEr CreateStreamingDistributionResponse)

DeleteCloudFrontOriginAccessIdentity

Delete an origin access identity.

See: DeleteCloudFrontOriginAccessIdentity

deleteCloudFrontOriginAccessIdentity

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dcfoaiId

-> State DeleteCloudFrontOriginAccessIdentity a

-> m DeleteCloudFrontOriginAccessIdentityResponse

deleteCloudFrontOriginAccessIdentityCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dcfoaiId

-> State DeleteCloudFrontOriginAccessIdentity a

-> m (Either ServiceEr DeleteCloudFrontOriginAccessIdentityResponse)

DeleteDistribution

Delete a distribution.

See: DeleteDistribution

deleteDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
ddId

-> State DeleteDistribution a

-> m DeleteDistributionResponse

deleteDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
ddId

-> State DeleteDistribution a

-> m (Either ServiceEr DeleteDistributionResponse)

DeleteStreamingDistribution

Delete a streaming distribution.

See: DeleteStreamingDistribution

deleteStreamingDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
dsdId

-> State DeleteStreamingDistribution a

-> m DeleteStreamingDistributionResponse

deleteStreamingDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
dsdId

-> State DeleteStreamingDistribution a

-> m (Either ServiceEr DeleteStreamingDistributionResponse)

GetCloudFrontOriginAccessIdentity

Get the information about an origin access identity.

See: GetCloudFrontOriginAccessIdentity

getCloudFrontOriginAccessIdentity

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gcfoaiId

-> State GetCloudFrontOriginAccessIdentity a

-> m GetCloudFrontOriginAccessIdentityResponse

getCloudFrontOriginAccessIdentityCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gcfoaiId

-> State GetCloudFrontOriginAccessIdentity a

-> m (Either ServiceEr GetCloudFrontOriginAccessIdentityResponse)

GetCloudFrontOriginAccessIdentityConfig

Get the configuration information about an origin access identity.

See: GetCloudFrontOriginAccessIdentityConfig

getCloudFrontOriginAccessIdentityConfig

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gcfoaicId

-> State GetCloudFrontOriginAccessIdentityConfig a

-> m GetCloudFrontOriginAccessIdentityConfigResponse

getCloudFrontOriginAccessIdentityConfigCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gcfoaicId

-> State GetCloudFrontOriginAccessIdentityConfig a

-> m (Either ServiceEr GetCloudFrontOriginAccessIdentityConfigResponse)

GetDistribution

Get the information about a distribution.

See: GetDistribution

getDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gdId

-> State GetDistribution a

-> m GetDistributionResponse

getDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gdId

-> State GetDistribution a

-> m (Either ServiceEr GetDistributionResponse)

GetDistributionConfig

Get the configuration information about a distribution.

See: GetDistributionConfig

getDistributionConfig

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gdcId

-> State GetDistributionConfig a

-> m GetDistributionConfigResponse

getDistributionConfigCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gdcId

-> State GetDistributionConfig a

-> m (Either ServiceEr GetDistributionConfigResponse)

GetInvalidation

Get the information about an invalidation.

See: GetInvalidation

getInvalidation

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
giDistributionId

-> Text
giId

-> State GetInvalidation a

-> m GetInvalidationResponse

getInvalidationCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
giDistributionId

-> Text
giId

-> State GetInvalidation a

-> m (Either ServiceEr GetInvalidationResponse)

GetStreamingDistribution

Get the information about a streaming distribution.

See: GetStreamingDistribution

getStreamingDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gsdId

-> State GetStreamingDistribution a

-> m GetStreamingDistributionResponse

getStreamingDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gsdId

-> State GetStreamingDistribution a

-> m (Either ServiceEr GetStreamingDistributionResponse)

GetStreamingDistributionConfig

Get the configuration information about a streaming distribution.

See: GetStreamingDistributionConfig

getStreamingDistributionConfig

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> Text
gsdcId

-> State GetStreamingDistributionConfig a

-> m GetStreamingDistributionConfigResponse

getStreamingDistributionConfigCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> Text
gsdcId

-> State GetStreamingDistributionConfig a

-> m (Either ServiceEr GetStreamingDistributionConfigResponse)

ListCloudFrontOriginAccessIdentities

List origin access identities.

See: ListCloudFrontOriginAccessIdentities

listCloudFrontOriginAccessIdentities :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListCloudFrontOriginAccessIdentities a -> ResumableSource m ListCloudFrontOriginAccessIdentitiesResponse

listCloudFrontOriginAccessIdentitiesCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListCloudFrontOriginAccessIdentities a -> ResumableSource m (Either ServiceEr ListCloudFrontOriginAccessIdentitiesResponse)

ListDistributions

List distributions.

See: ListDistributions

listDistributions :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListDistributions a -> ResumableSource m ListDistributionsResponse

listDistributionsCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListDistributions a -> ResumableSource m (Either ServiceEr ListDistributionsResponse)

ListInvalidations

List invalidation batches.

See: ListInvalidations

listInvalidations

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> Text
liDistributionId

-> State ListInvalidations a

-> ResumableSource m ListInvalidationsResponse

listInvalidationsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> Text
liDistributionId

-> State ListInvalidations a

-> ResumableSource m (Either ServiceEr ListInvalidationsResponse)

ListStreamingDistributions

List streaming distributions.

See: ListStreamingDistributions

listStreamingDistributions :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m)) => State ListStreamingDistributions a -> ResumableSource m ListStreamingDistributionsResponse

listStreamingDistributionsCatch :: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m)) => State ListStreamingDistributions a -> ResumableSource m (Either ServiceEr ListStreamingDistributionsResponse)

UpdateCloudFrontOriginAccessIdentity

Update an origin access identity.

See: UpdateCloudFrontOriginAccessIdentity

updateCloudFrontOriginAccessIdentity

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> CloudFrontOriginAccessIdentityConfig
ucfoaiCloudFrontOriginAccessIdentityConfig

-> Text
ucfoaiId

-> State UpdateCloudFrontOriginAccessIdentity a

-> m UpdateCloudFrontOriginAccessIdentityResponse

updateCloudFrontOriginAccessIdentityCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> CloudFrontOriginAccessIdentityConfig
ucfoaiCloudFrontOriginAccessIdentityConfig

-> Text
ucfoaiId

-> State UpdateCloudFrontOriginAccessIdentity a

-> m (Either ServiceEr UpdateCloudFrontOriginAccessIdentityResponse)

UpdateDistribution

Update a distribution.

See: UpdateDistribution

updateDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> DistributionConfig
udDistributionConfig

-> Text
udId

-> State UpdateDistribution a

-> m UpdateDistributionResponse

updateDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> DistributionConfig
udDistributionConfig

-> Text
udId

-> State UpdateDistribution a

-> m (Either ServiceEr UpdateDistributionResponse)

UpdateStreamingDistribution

Update a streaming distribution.

See: UpdateStreamingDistribution

updateStreamingDistribution

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> StreamingDistributionConfig
usdStreamingDistributionConfig

-> Text
usdId

-> State UpdateStreamingDistribution a

-> m UpdateStreamingDistributionResponse

updateStreamingDistributionCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> StreamingDistributionConfig
usdStreamingDistributionConfig

-> Text
usdId

-> State UpdateStreamingDistribution a

-> m (Either ServiceEr UpdateStreamingDistributionResponse)

Re-exported

module Network.AWS.CloudFront.V2014_05_31

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()