Network.AWS.CloudFront.ListCloudFrontOriginAccessIdentities

Description

Lists origin access identities.

This operation returns paginated results.

Creating a Request

listCloudFrontOriginAccessIdentities :: ListCloudFrontOriginAccessIdentities

Creates a value of ListCloudFrontOriginAccessIdentities with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcfoaiMarker - Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

lcfoaiMaxItems - The maximum number of origin access identities you want in the response body.

data ListCloudFrontOriginAccessIdentities

The request to list origin access identities.

See: listCloudFrontOriginAccessIdentities smart constructor.

Request Lenses

lcfoaiMarker :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

lcfoaiMaxItems :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

The maximum number of origin access identities you want in the response body.

Destructuring the Response

listCloudFrontOriginAccessIdentitiesResponse

Arguments

:: Int
lcfoairsResponseStatus

-> CloudFrontOriginAccessIdentityList
lcfoairsCloudFrontOriginAccessIdentityList

-> ListCloudFrontOriginAccessIdentitiesResponse

Creates a value of ListCloudFrontOriginAccessIdentitiesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcfoairsResponseStatus - -- | The response status code.

lcfoairsCloudFrontOriginAccessIdentityList - The CloudFrontOriginAccessIdentityList type.

data ListCloudFrontOriginAccessIdentitiesResponse

The returned result of the corresponding request.

See: listCloudFrontOriginAccessIdentitiesResponse smart constructor.

Response Lenses

lcfoairsResponseStatus :: Lens' ListCloudFrontOriginAccessIdentitiesResponse Int

- | The response status code.

lcfoairsCloudFrontOriginAccessIdentityList :: Lens' ListCloudFrontOriginAccessIdentitiesResponse CloudFrontOriginAccessIdentityList

The CloudFrontOriginAccessIdentityList type.