Network.AWS.CloudFront.ListCloudFrontOriginAccessIdentities

Description

List origin access identities.

http://docs.aws.amazon.com/AmazonCloudFront/latest/APIReference/ListCloudFrontOriginAccessIdentities.html

Request

data ListCloudFrontOriginAccessIdentities

Request constructor

listCloudFrontOriginAccessIdentities :: ListCloudFrontOriginAccessIdentities

ListCloudFrontOriginAccessIdentities constructor.

The fields accessible through corresponding lenses are:

lcfoaiMarker :: Maybe Text

lcfoaiMaxItems :: Maybe Text

Request lenses

lcfoaiMarker :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

lcfoaiMaxItems :: Lens' ListCloudFrontOriginAccessIdentities (Maybe Text)

The maximum number of origin access identities you want in the response body.

Response

data ListCloudFrontOriginAccessIdentitiesResponse

Response constructor

listCloudFrontOriginAccessIdentitiesResponse

Arguments

:: CloudFrontOriginAccessIdentityList
lcfoairCloudFrontOriginAccessIdentityList

-> ListCloudFrontOriginAccessIdentitiesResponse

ListCloudFrontOriginAccessIdentitiesResponse constructor.

The fields accessible through corresponding lenses are:

lcfoairCloudFrontOriginAccessIdentityList :: CloudFrontOriginAccessIdentityList

Response lenses

lcfoairCloudFrontOriginAccessIdentityList :: Lens' ListCloudFrontOriginAccessIdentitiesResponse CloudFrontOriginAccessIdentityList

The CloudFrontOriginAccessIdentityList type.