Network.AWS.CloudFront.ListDistributions

Description

List distributions.

This operation returns paginated results.

Creating a Request

listDistributions :: ListDistributions

Creates a value of ListDistributions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldMarker - Use this when paginating results to indicate where to begin in your list of distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

ldMaxItems - The maximum number of distributions you want in the response body.

data ListDistributions

The request to list your distributions.

See: listDistributions smart constructor.

Request Lenses

ldMarker :: Lens' ListDistributions (Maybe Text)

Use this when paginating results to indicate where to begin in your list of distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

ldMaxItems :: Lens' ListDistributions (Maybe Text)

The maximum number of distributions you want in the response body.

Destructuring the Response

listDistributionsResponse

Arguments

:: Int
ldrsResponseStatus

-> DistributionList
ldrsDistributionList

-> ListDistributionsResponse

Creates a value of ListDistributionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrsResponseStatus - -- | The response status code.

ldrsDistributionList - The DistributionList type.

data ListDistributionsResponse

The returned result of the corresponding request.

See: listDistributionsResponse smart constructor.

Response Lenses

ldrsResponseStatus :: Lens' ListDistributionsResponse Int

- | The response status code.

ldrsDistributionList :: Lens' ListDistributionsResponse DistributionList

The DistributionList type.