Network.AWS.CloudFront.ListInvalidations

Description

Lists invalidation batches.

This operation returns paginated results.

Creating a Request

listInvalidations

Arguments

:: Text
liDistributionId

-> ListInvalidations

Creates a value of ListInvalidations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

liMarker - Use this parameter when paginating results to indicate where to begin in your list of invalidation batches. Because the results are returned in decreasing order from most recent to oldest, the most recent results are on the first page, the second page will contain earlier results, and so on. To get the next page of results, set Marker to the value of the NextMarker from the current page's response. This value is the same as the ID of the last invalidation batch on that page.

liMaxItems - The maximum number of invalidation batches that you want in the response body.

liDistributionId - The distribution's ID.

data ListInvalidations

The request to list invalidations.

See: listInvalidations smart constructor.

Request Lenses

liMarker :: Lens' ListInvalidations (Maybe Text)

Use this parameter when paginating results to indicate where to begin in your list of invalidation batches. Because the results are returned in decreasing order from most recent to oldest, the most recent results are on the first page, the second page will contain earlier results, and so on. To get the next page of results, set Marker to the value of the NextMarker from the current page's response. This value is the same as the ID of the last invalidation batch on that page.

liMaxItems :: Lens' ListInvalidations (Maybe Text)

The maximum number of invalidation batches that you want in the response body.

liDistributionId :: Lens' ListInvalidations Text

The distribution's ID.

Destructuring the Response

listInvalidationsResponse

Arguments

:: Int
lirsResponseStatus

-> InvalidationList
lirsInvalidationList

-> ListInvalidationsResponse

Creates a value of ListInvalidationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lirsResponseStatus - -- | The response status code.

lirsInvalidationList - Information about invalidation batches.

data ListInvalidationsResponse

The returned result of the corresponding request.

See: listInvalidationsResponse smart constructor.

Response Lenses

lirsResponseStatus :: Lens' ListInvalidationsResponse Int

- | The response status code.

lirsInvalidationList :: Lens' ListInvalidationsResponse InvalidationList

Information about invalidation batches.