Network.AWS.CloudFront.ListStreamingDistributions

Description

List streaming distributions.

This operation returns paginated results.

Creating a Request

listStreamingDistributions :: ListStreamingDistributions

Creates a value of ListStreamingDistributions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsdMarker - The value that you provided for the Marker request parameter.

lsdMaxItems - The value that you provided for the MaxItems request parameter.

data ListStreamingDistributions

The request to list your streaming distributions.

See: listStreamingDistributions smart constructor.

Request Lenses

lsdMarker :: Lens' ListStreamingDistributions (Maybe Text)

The value that you provided for the Marker request parameter.

lsdMaxItems :: Lens' ListStreamingDistributions (Maybe Text)

The value that you provided for the MaxItems request parameter.

Destructuring the Response

listStreamingDistributionsResponse

Arguments

:: Int
lsdrsResponseStatus

-> StreamingDistributionList
lsdrsStreamingDistributionList

-> ListStreamingDistributionsResponse

Creates a value of ListStreamingDistributionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsdrsResponseStatus - -- | The response status code.

lsdrsStreamingDistributionList - The StreamingDistributionList type.

data ListStreamingDistributionsResponse

The returned result of the corresponding request.

See: listStreamingDistributionsResponse smart constructor.

Response Lenses

lsdrsResponseStatus :: Lens' ListStreamingDistributionsResponse Int

- | The response status code.

lsdrsStreamingDistributionList :: Lens' ListStreamingDistributionsResponse StreamingDistributionList

The StreamingDistributionList type.