Network.AWS.CloudFront.V2014_05_31.ListStreamingDistributions

Description

List streaming distributions.

Documentation

listStreamingDistributions :: ListStreamingDistributions

Minimum specification for a ListStreamingDistributions request.

data ListStreamingDistributions

Constructors

ListStreamingDistributions

Fields

_lsdrMarker :: Maybe Text

Use this when paginating results to indicate where to begin in your list of streaming distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

_lsdrMaxItems :: Maybe Text

The maximum number of streaming distributions you want in the response body.

lsdrMaxItems :: Lens' ListStreamingDistributions (Maybe Text)

lsdrMarker :: Lens' ListStreamingDistributions (Maybe Text)

data ListStreamingDistributionsResponse

Constructors

ListStreamingDistributionsResponse

Fields

_lsdsStreamingDistributionList :: StreamingDistributionList

The StreamingDistributionList type.

lsdsStreamingDistributionList :: Iso' ListStreamingDistributionsResponse StreamingDistributionList